==> 02.descriptive.Rmd <==
---
title: "Four Week Update"
author: "Andie Creel"
date: "7/1/2021"
output: 
  pdf_document:
    number_sections: true
    extra_dependencies: ["float"]

knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, encoding = encoding, output_file = file.path(dirname(inputFile), 'Exploratory_Output/regressions/four_week_update.pdf')) })
---

```{r, include=FALSE}
#R Markdown stuff
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(fig.pos = "!H", out.extra = "")
options(tinytex.verbose = TRUE)
```

```{r setup, include=FALSE}
library(vroom)
library(tidyr)
library(stringr)
library(usmap)
library(tidycensus)
library(ggplot2)
library(fixest)
library(kableExtra)
library(gtools)
library(forcats)
library("wesanderson")
options(scipen=99999)
library(janitor)
library(dplyr)
# ---- #
library(choroplethr)
library(choroplethrMaps)
library(usdata)
library(maps)
data(fips_codes)
data(continental_us_states)
# --- #



myWorking <- vroom("Datasets/clean_data/lwcf_annualPop_deccenialDemographics.csv")

# additional data cleaning ------------

# Removing counties whose FIPSs code I didn't identify 
myWorking <- myWorking %>%
  filter(!is.na(fips)) %>%
  distinct() %>%
  mutate(type = as.factor(type)) %>%
  mutate(state_fips = as.factor(str_sub(fips, 1, 2))) %>%
  filter(!is.na(white_pct)) %>%
  dplyr::mutate(POC_pct = 100 - white_pct)

# Getting quantiles (Sims 2020)

#Getting the quantile for each state each year! 
myQuants <- myWorking %>%
  dplyr::group_by(state_fips, year) %>%
  dplyr::summarise(quants = list(quantile(POC_pct, probs = c(.25, .5, .75)))) %>%
  unnest_wider(quants) %>%
  dplyr::rename(poc_25 = `25%`) %>%
  dplyr::rename(poc_50 = `50%`) %>%
  dplyr::rename(poc_75 = `75%`)

#joining back in
myWorking <- left_join(myWorking, myQuants, by = c("state_fips", "year"))
rm(myQuants)

#making a factor variable for the quantiles (by state and year)
myWorking <- myWorking %>%
  mutate(poc_quants = if_else(POC_pct > poc_75, "4", "")) %>% # fourth quantile 
  mutate(poc_quants = if_else(POC_pct <= poc_75, "3", poc_quants)) %>% # third quantile
  mutate(poc_quants = if_else(POC_pct <= poc_50, "2", poc_quants)) %>% # second quantile
  mutate(poc_quants = if_else(POC_pct <= poc_25, "1", poc_quants)) %>% # first quantile 
  select(-poc_75, -poc_50, - poc_25) %>%
  mutate(poc_quants = as.factor(poc_quants))



myLogLin <- function(x){
  y <- (exp(x)-1)*100 
  return(y)
}


```


```{r DATA-MANIPULATION, echo=FALSE, message=FALSE}

# --- data manipulation to get quantity of grants per quantile--- #

# total grants per state 
myTotal <- myWorking %>%
  group_by(state_fips) %>%
  mutate(state_total = sum(got_grant)) %>%
  select(state_fips, state_total) %>%
  distinct()

# number of grants given to first quant counties ever per state
myOne <- myWorking %>%
  filter(poc_quants == "1") %>%
  group_by(state_fips) %>%
  mutate(one_total = sum(got_grant)) %>%
  select(state_fips, one_total) %>%
  distinct()

# number of grants given to second quant counties ever per state
myTwo <- myWorking %>%
  filter(poc_quants == "2") %>%
  group_by(state_fips) %>%
  mutate(two_total = sum(got_grant)) %>%
  select(state_fips, two_total) %>%
  distinct()

# number of grants given to third quant counties ever per state
myThree <- myWorking %>%
  filter(poc_quants == "3") %>%
  group_by(state_fips) %>%
  mutate(three_total = sum(got_grant)) %>%
  select(state_fips, three_total) %>%
  distinct()

# number of grants given to third quant counties ever per state
myFour <- myWorking %>%
  filter(poc_quants == "4") %>%
  group_by(state_fips) %>%
  mutate(four_total = sum(got_grant)) %>%
  select(state_fips, four_total) %>%
  distinct()

#merge
myPOC_mapData <- left_join(myTotal, myOne , by = "state_fips")
myPOC_mapData <- left_join(myPOC_mapData, myTwo , by = "state_fips")
myPOC_mapData <- left_join(myPOC_mapData, myThree , by = "state_fips")
myPOC_mapData <- left_join(myPOC_mapData, myFour , by = "state_fips")
rm(myTotal, myOne, myTwo, myThree, myFour)

#percents 
myPOC_mapData <- myPOC_mapData %>%
  mutate(one_pct = round(one_total/state_total * 100, 0 )) %>%
  mutate(two_pct = round(two_total/state_total * 100, 0 )) %>%
  mutate(three_pct = round(three_total/state_total * 100 , 0)) %>%
  mutate(four_pct = round(four_total/state_total * 100, 0 )) 
```


# Choropleth Maps: POC Quartiles

I calculate the percentage a county's population that are people of color (POC) by taking 100 - % of people who are non-hispanic white. 

I then categorize counties into four quartiles based on the % of the population that is POC. The first quartile categorizes the whitest counties and the fourth quartile categorizes the couties with the most POC. 

When calculating quartiles, I group by state and decade. I group by state because the LWCF is administered at the state level, therefore managers are making **intrastate** comparisons when deciding how to allocate funds. I group by decade because the % of the population that is POC changes through time. Therefore, a county that had a significant number of POC in the 1970s may have experienced demographic changes by the 2010s. I group at the decade level because my demographic characteristics come from the decennial census which is collected at the decade level.  


## Percent of Grants Given to Each Quartile 

Four each of the four quartiles, I calculate the percentage received of the total grants given from 1966 until 2018. This metric is a helpful moment for the total quantity of grants given to the different quartiles, but ignores time trends as well as the amount of funding allocated. 

```{r, echo=FALSE, message=FALSE, warning=FALSE}

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myPOC_mapData <- left_join(myPOC_mapData, fips_codes, by = c("state_fips" = "state_code")) 

myPOC_mapData <- myPOC_mapData %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region))


# exporting map to make tableau 
# vroom_write(myPOC_mapData, "Datasets/clean_data/mapPOCquants.csv", delim = ",")

# --- maps --- #

# First Quantile
myOne_map <- myPOC_mapData %>%
  select(region, one_pct) %>%
  rename(value = one_pct) %>%
  filter(!is.na(region))

state_choropleth(myOne_map,
  title = "% of grants given to first quantile of POC counties",
  legend = "Percent %",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr")+
  ggtitle(label = "% of total grants given to first quantile POC counties", subtitle =  "(whitest counties)") + 
  labs(fill = "% received by 1st")

# Second Quantile
myTwo_map <- myPOC_mapData %>%
  select(region, two_pct) %>%
  rename(value = two_pct) %>%
  filter(!is.na(region))

state_choropleth(myTwo_map,
  title = "% of grants given to second quantile of POC counties",
  legend = "Percent %",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "% received by 2nd")


# Third Quantile
myThree_map <- myPOC_mapData %>%
  select(region, three_pct) %>%
  rename(value = three_pct) %>%
  filter(!is.na(region))

state_choropleth(myThree_map,
  title = "% of grants given to third quantile of POC counties",
  legend = "Percent %",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr")+ 
  labs(fill = "% received by 3rd")

# Four Quantile
myFour_map <- myPOC_mapData %>%
  select(region, four_pct) %>%
  rename(value = four_pct) %>%
  filter(!is.na(region))

state_choropleth(myFour_map,
  title = "% of grants given to fourth quantile of POC counties",
  legend = "Percent %",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "% received by 4th") + 
    ggtitle(label = "% of grants given to fourth quantile of POC counties", subtitle =  "(counties with most POC)") 

# difference between fourth and first quantile 

myDifference_map <- myPOC_mapData %>%
  select(region, one_pct, four_pct) %>%
  mutate(value = round((four_pct - one_pct)/ four_pct * 100)) %>%
  filter(!is.na(region))

state_choropleth(myDifference_map,
  title = "% difference in quantity grants recieved by fourth quartile and first",
  legend = "Percent %",
  zoom = continental_us_states,
  num_colors = 7) +
  scale_fill_brewer(palette="PuOr") + 
  labs(fill = "% diff. between 4th and 1st")

ggsave(width = 11, height = 8, filename = "Exploratory_Output/maps/difference_POC_quants.png")



```

## Average Grant Size per Capita

I calculate the per capita amount of grant funding for each year and include years when the amount is zero. I find the average per capita amount of grant funding for each quartile. 


```{r DATA-MANIPULATION-2, echo=FALSE, message=FALSE, warning= FALSE}

# --- data manipulation to get amount per person --- #

#getting region in here for the state_choropleth function
myWorking <- left_join(myWorking, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region))

# calculating grant award per capita for each county each year
myWorking <- myWorking %>%
  mutate(amount_perCap = if_else(!is.na(amount), amount / population,  0))

# -- first quartile maps -- #

# First quartile 
myPerCap_map1 <- myWorking %>%
  filter(poc_quants == "1") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)
  

state_choropleth(myPerCap_map1,
  title = "Average per capita grants amount given to first quantile of POC counties",
  legend = "Average per cap. $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr")+
  ggtitle(label = "Average per capita grants amount given to first quantile of POC counties", 
          subtitle =  "(whitest counties)") + 
  labs(fill = "Average per cap. $")

#second quartile 
myPerCap_map2 <- myWorking %>%
  filter(poc_quants == "2") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)
  

state_choropleth(myPerCap_map2,
  title = "Average per capita grants amount given to second quantile of POC counties",
  legend = "Average per cap. $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "Average per cap. $")

  
#third quartile 
myPerCap_map3 <- myWorking %>%
  filter(poc_quants == "3") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)
  

state_choropleth(myPerCap_map3,
  title = "Average per capita grants amount given to thrid quantile of POC counties",
  legend = "Average per cap. $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "Average per cap. $")

#fourth quartile 
myPerCap_map4 <- myWorking %>%
  filter(poc_quants == "4") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)
  
state_choropleth(myPerCap_map4,
  title = "Average per capita grants amount given to fourth quantile of POC counties",
  legend = "Average per cap. $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "Average per cap. $")


# -- difference -- #


myPerCap_difference_map <- left_join(x = myPerCap_map4, y = myPerCap_map1, by = "region") %>%
  rename(value_4 = value.x) %>%
  rename(value_1 = value.y) %>%
  mutate(value = round((value_4 - value_1)/ value_4 * 100)) %>%
  filter(!is.na(region))

state_choropleth(myPerCap_difference_map,
  title = "% difference in avg. per capita grant between fourth quartile and first",
  legend = "% difference",
  zoom = continental_us_states,
  num_colors = 7) +
  scale_fill_brewer(palette="YlOrBr", direction=-1) + 
  labs(fill = "% diff. between 4th and 1st")

ggsave(width = 11, height = 8, filename = "Exploratory_Output/maps/difference_POC_quants.png")



```



















<!-- # Whether or not a county gets a grant -->
<!-- Zeros are back in -->

```{r, include = FALSE, echo=FALSE, message=FALSE, eval = FALSE}
# Exploratory work
# 
# # Median income 
# white_pct <- glm(family = "binomial", data = myWorking, formula = got_grant ~ white_pct )
# black_pct <- glm(family = "binomial", data = myWorking, formula = got_grant ~ black_pct)
# 
# summary(white_pct)
# exp(coef(white_pct))
# 
# summary(black_pct)
# exp(coef(black_pct))
# 
# white_four <- glm(family = "binomial", data = myWorking, formula = got_grant ~ white_four )
# black_four <- glm(family = "binomial", data = myWorking, formula = got_grant ~ black_four)
# 
# summary(white_four)
# exp(coef(white_four))
# 
# summary(black_four)
# exp(coef(black_four))
# 


summary(glm(family = "binomial", data = myWorking, formula = got_grant ~ as.factor(poc_quants)))
femlm(family = "logit", data = myWorking, fml = got_grant ~ POC_pct | state_fips)
femlm(family = "logit", data = myWorking, fml = got_grant ~ POC_pct)


```


```{r, echo=FALSE, message=FALSE, include=FALSE, eval = FALSE}
#Replicating UCLA

myTry <- myWorking %>%
  dplyr::select(got_grant, white_pct, med_income_house) %>%
  filter(!is.na(white_pct)) %>%
  dplyr::mutate(POC_pct = 100 - white_pct)

#Summary stats
summary(myTry)
sapply(myTry, sd)


# --- SIMPLE REGRESSION ON PERCENT POC --- #
myPOC <- glm(got_grant ~ POC_pct, data = myTry, family = "binomial")
summary(myPOC)
confint(myPOC)

newData_poc <- myTry %>%
  dplyr::select(POC_pct)

newData_poc$pocP <- predict(myPOC, newdata = newData_poc, type = "response")


newdata2_poc <- cbind(newData_poc, predict(myPOC, newdata = newData_poc, type = "link",
    se = TRUE))

newdata2_poc <- within(newdata2_poc, {
    PredictedProb <- plogis(fit)
    LL <- plogis(fit - (1.96 * se.fit))
    UL <- plogis(fit + (1.96 * se.fit))
})


ggplot(newdata2_poc, aes(x = POC_pct, y = PredictedProb)) + geom_ribbon(aes(ymin = LL,
    ymax = UL), alpha = 0.2) + geom_line(size = 1)


# --- SIMPLE REGRESSION ON MEDIAN HOUSEHOLD INCOME --- # 


myIncome <- glm(got_grant ~ med_income_house, data = myTry, family = "binomial")
summary(myIncome)
confint(myIncome)

newData_inc <- myTry %>%
  dplyr::select(med_income_house)

newData_inc$incP <- predict(myIncome, newdata = newData_inc, type = "response")

newdata2_inc <- cbind(newData_inc, predict(myIncome, newdata = newData_inc, type = "link",
    se = TRUE))

newdata2_inc <- within(newdata2_inc, {
    PredictedProb <- plogis(fit)
    LL <- plogis(fit - (1.96 * se.fit))
    UL <- plogis(fit + (1.96 * se.fit))
})


ggplot(newdata2_inc, aes(x = med_income_house, y = PredictedProb)) + geom_ribbon(aes(ymin = LL,
  ymax = UL), alpha = 0.2) + geom_line(size = 1)



```

<!-- Next steps  -->
<!-- * Control for income when looking at percentage of POC, because something odd is happening -->
<!-- * Could also control for home ownership, education, poverty etc (going to discuss with Sims at noon) -->
