==> 02.exploratory_regressions.Rmd <==
---
title: "Exploratory Regressions"
author: "Andie Creel"
date: "6/15/2021"
output: 
  pdf_document:
    extra_dependencies: ["float"]

knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, encoding = encoding, output_file = file.path(dirname(inputFile), 'Exploratory_Output/regressions/exploratory_regressions.pdf')) })
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(fig.pos = "!H", out.extra = "")


library(vroom)
library(dplyr)
library(tidyr)
library(stringr)
library(usmap)
library(tidycensus)
library(ggplot2)
library(fixest)
library(kableExtra)
library("wesanderson")
options(scipen=99999)
library(gtools)
library(forcats)

myWorking <- vroom("Datasets/clean_data/lwcf_tws_projects_IPUMS&ACS.csv")

# Removing useless variables 
myWorking <- myWorking %>%
  dplyr::select(-c(objectid, county, grantidelement, grantelementtitle, grantsponsor, congdistrict, relate, starts_with("urban"), starts_with("rural")))

# Removing counties whose FIPSs code I didn't identify 
myWorking <- myWorking %>%
  filter(!is.na(county_fips))

```

# What's the data look like

```{r, echo= FALSE, fig.width=3, fig.height=2, warning=FALSE}
# distribution of amount
ggplot(myWorking, aes(x=amount)) + 
  geom_histogram(binwidth=100000) +
  theme_minimal() +
  ggtitle("Count of projects' award size")

# distribution of log amount
ggplot(myWorking, aes(x=log(amount))) + 
  geom_histogram(binwidth=.1) +
  theme_minimal()+
  ggtitle("Count of LOG of projects' award size")

# 
ggplot(myWorking, aes(x=fiscal_year)) + 
  geom_histogram(binwidth=1) +
  theme_minimal()+
  ggtitle("Count of awards by year") +
  scale_x_continuous(breaks=seq(1965,2020,10))



```


## Types of LWCF projects

Types of LWCF projects 

* A: Land Acquisition
* D: Development 
* R: Redevelopment 
* C: Combination
* P: Planning

```{r, echo= FALSE, fig.width=3.5, fig.height=2.5, warning=FALSE}

# distribution of  amount
ggplot(myWorking, aes(x=amount, fill=type, color = type)) + 
  geom_histogram(binwidth=10000) +
  theme_minimal()+
  ggtitle("Count of projects' award size") 

# # truncated distribution of  amount
# ggplot(myWorking, aes(x=amount, fill=type, color = type)) + 
#   geom_histogram(binwidth=10000) +
#   theme_minimal()+
#   ggtitle("Count of projects' award size: truncated") +
#   xlim(0,500000)


# distribution of log amount
ggplot(myWorking, aes(x=log(amount), fill=type, color = type)) + 
  geom_histogram(binwidth=.1) +
  theme_minimal()+
  ggtitle("Count of LOG of projects' award size")

# 
ggplot(myWorking, aes(x=fiscal_year, fill=type, color = type)) + 
  geom_histogram(binwidth=1) +
  theme_minimal()+
  ggtitle("Count of awards by year") +
  scale_x_continuous(breaks=seq(1965,2020,10))

```

# Regressions

## Outcome of interest: log of size of individual grant awards

### Socioeconomic 

```{r,  echo=FALSE, message=FALSE}

# Initial regressions -------------------

myWorking <- myWorking %>%
  mutate(type = as.factor(type))

#park type control
type <- feols(data = myWorking, log(amount) ~ type + population| state_fips )

#economic 
poverty <- feols(data = myWorking, log(amount) ~ inc_below_pov_pct + type + population| state_fips )
med_inc <- feols(data = myWorking, log(amount) ~ med_income_house + type + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 
```

The percentage of the county that is below the poverty line has a large, significant, negative effect on the size of the grant awarded to that county. Possible explanation: the need to have matching funds for state and local programs, so poorer communities raise less for matching? 


### Race

```{r,  echo=FALSE, message=FALSE}

# race
white <- feols(data = myWorking, log(amount) ~ white_pct + type + population| state_fips )
black <- feols(data = myWorking, log(amount) ~ black_pct + type + population| state_fips )
native <- feols(data = myWorking, log(amount) ~ native_pct + type + population| state_fips )
hispanic <- feols(data = myWorking, log(amount) ~ hispanic_pct + type + population| state_fips )
asian <- feols(data = myWorking, log(amount) ~ asian_pct + type + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# resids <- residuals(one)
```


The percentage of the county that is black has a large, significant, **positive** effect on the size of the grant awarded to that county. The percent of the county that is white a significant, **negative** effect on size of grant awarded in that county. Other races and ethnicities are not precisely estimated. Possible explanation: the need to have matching funds for state and local programs, effective community organizing (test the effect size post civil rights movement)? 


## Before and after 1980



```{r,  echo=FALSE, message=FALSE}
myWorking_pre80 <- myWorking %>%
  filter(fiscal_year <= 1980)

myWorking_post80 <- myWorking %>%
  filter(fiscal_year > 1980)



# income -----------

#park type control
type <- feols(data = myWorking_pre80, log(amount) ~ type + population| state_fips )

#economic 
poverty <- feols(data = myWorking_pre80, log(amount) ~ inc_below_pov_pct + type + population| state_fips )
med_inc <- feols(data = myWorking_pre80, log(amount) ~ med_income_house + type + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income: pre 1980") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 


#park type control
type <- feols(data = myWorking_post80, log(amount) ~ type + population| state_fips )

#economic 
poverty <- feols(data = myWorking_post80, log(amount) ~ inc_below_pov_pct + type + population| state_fips )
med_inc <- feols(data = myWorking_post80, log(amount) ~ med_income_house + type + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income: post 80") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 





# race --------------

white <- feols(data = myWorking_pre80, log(amount) ~ white_pct + type + population| state_fips )
black <- feols(data = myWorking_pre80, log(amount) ~ black_pct + type + population| state_fips )
native <- feols(data = myWorking_pre80, log(amount) ~ native_pct + type + population| state_fips )
hispanic <- feols(data = myWorking_pre80, log(amount) ~ hispanic_pct + type + population| state_fips )
asian <- feols(data = myWorking_pre80, log(amount) ~ asian_pct + type + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race: pre 80") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 



white <- feols(data = myWorking_post80, log(amount) ~ white_pct + type + population| state_fips )
black <- feols(data = myWorking_post80, log(amount) ~ black_pct + type + population| state_fips )
native <- feols(data = myWorking_post80, log(amount) ~ native_pct + type + population| state_fips )
hispanic <- feols(data = myWorking_post80, log(amount) ~ hispanic_pct + type + population| state_fips )
asian <- feols(data = myWorking_post80, log(amount) ~ asian_pct + type + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct", 
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race: post 80") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 



```

* There don't seem to be major differences in the effect of poverty before and after 1980. 

* The effect of an addition percent of the community being white (or black) doesn't change that much 

* The effect for an additional pct of the county being asian is noisy but may be worth trying to pull apart

* The effect for an additianal pct of the county being hispanic is precesly estimated (and large and postive) after 1980 but not prior


# Grant Type

* A: Land Acquisition
* D: Development 
* R: Redevelopment 
* C: Combination
* P: Planning -- The planning grants don't have a county (the county field is 'planning'). They appear to be awarded at the state level, for the most part. 


```{r scrap, echo=FALSE, message=FALSE}

myWorking <- myWorking %>%
  mutate(type = as.character(type)) %>%
  mutate(type = str_sub(type, 1, 1))

myWorking_A <- myWorking %>%
  dplyr::filter(type == "A")

myWorking_D <- myWorking %>%
  dplyr::filter(type == "D")

myWorking_R <- myWorking %>%
  dplyr::filter(type == "R")

myWorking_C <- myWorking %>%
  dplyr::filter(type == "C")

myWorking_P <- myWorking %>%
  dplyr::filter(type == "P")

```

## Socio-economic

### Land Acquisition 

```{r,  echo=FALSE, message=FALSE}

# Initial regressions -------------------

myWorking_A <- myWorking_A %>%
  mutate(type = as.factor(type))

#park type control
type <- feols(data = myWorking_A, log(amount) ~ population| state_fips )

#economic 
poverty <- feols(data = myWorking_A, log(amount) ~ inc_below_pov_pct  + population| state_fips )
med_inc <- feols(data = myWorking_A, log(amount) ~ med_income_house  + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 


```

### Development 

```{r,  echo=FALSE, message=FALSE}

# Initial regressions -------------------

myWorking_D <- myWorking_D %>%
  mutate(type = as.factor(type))

#park type control
type <- feols(data = myWorking_D, log(amount) ~ population| state_fips )

#economic 
poverty <- feols(data = myWorking_D, log(amount) ~ inc_below_pov_pct  + population| state_fips )
med_inc <- feols(data = myWorking_D, log(amount) ~ med_income_house  + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 



```

### Redevlopment

```{r,  echo=FALSE, message=FALSE}

# Initial regressions -------------------

myWorking_R <- myWorking_R %>%
  mutate(type = as.factor(type))

#park type control
type <- feols(data = myWorking_R, log(amount) ~ population| state_fips )

#economic 
poverty <- feols(data = myWorking_R, log(amount) ~ inc_below_pov_pct  + population| state_fips )
med_inc <- feols(data = myWorking_R, log(amount) ~ med_income_house  + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

```

### Combination 

```{r,  echo=FALSE, message=FALSE}

# Initial regressions -------------------

myWorking_C <- myWorking_C %>%
  mutate(type = as.factor(type))

#park type control
type <- feols(data = myWorking_C, log(amount) ~ population| state_fips )

#economic 
poverty <- feols(data = myWorking_C, log(amount) ~ inc_below_pov_pct  + population| state_fips )
med_inc <- feols(data = myWorking_C, log(amount) ~ med_income_house  + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 


```


## Race

### Land Acquisition 

```{r,  echo=FALSE, message=FALSE}

# race
white <- feols(data = myWorking_A, log(amount) ~ white_pct  + population| state_fips )
black <- feols(data = myWorking_A, log(amount) ~ black_pct  + population| state_fips )
native <- feols(data = myWorking_A, log(amount) ~ native_pct  + population| state_fips )
hispanic <- feols(data = myWorking_A, log(amount) ~ hispanic_pct  + population| state_fips )
asian <- feols(data = myWorking_A, log(amount) ~ asian_pct  + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# resids <- residuals(one)
```

### Development 

```{r,  echo=FALSE, message=FALSE}

# race
white <- feols(data = myWorking_D, log(amount) ~ white_pct  + population| state_fips )
black <- feols(data = myWorking_D, log(amount) ~ black_pct  + population| state_fips )
native <- feols(data = myWorking_D, log(amount) ~ native_pct  + population| state_fips )
hispanic <- feols(data = myWorking_D, log(amount) ~ hispanic_pct  + population| state_fips )
asian <- feols(data = myWorking_D, log(amount) ~ asian_pct  + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 


```


### Redevelopment

```{r,  echo=FALSE, message=FALSE}

# race
white <- feols(data = myWorking_R, log(amount) ~ white_pct  + population| state_fips )
black <- feols(data = myWorking_R, log(amount) ~ black_pct  + population| state_fips )
native <- feols(data = myWorking_R, log(amount) ~ native_pct  + population| state_fips )
hispanic <- feols(data = myWorking_R, log(amount) ~ hispanic_pct  + population| state_fips )
asian <- feols(data = myWorking_R, log(amount) ~ asian_pct  + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

```


### Combination 

```{r,  echo=FALSE, message=FALSE}

# race
white <- feols(data = myWorking_C, log(amount) ~ white_pct  + population| state_fips )
black <- feols(data = myWorking_C, log(amount) ~ black_pct  + population| state_fips )
native <- feols(data = myWorking_C, log(amount) ~ native_pct  + population| state_fips )
hispanic <- feols(data = myWorking_C, log(amount) ~ hispanic_pct  + population| state_fips )
asian <- feols(data = myWorking_C, log(amount) ~ asian_pct  + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

```


# Quartiles for Demographics 

```{r,  echo=FALSE, message=FALSE}
# creating 5 factor variables for racial groups and poverty 

myWorking <- myWorking %>%
  mutate(black_quant = quantcut(black_pct, q = 5, na.rm=TRUE)) %>%
  mutate(white_quant = quantcut(white_pct, q = 5, na.rm=TRUE)) %>%
  mutate(asian_quant = quantcut(asian_pct, q = 5, na.rm=TRUE)) %>%
  mutate(native_quant = quantcut(native_pct, q = 5, na.rm=TRUE)) %>%
  mutate(hispanic_quant = quantcut(hispanic_pct, q = 5, na.rm=TRUE)) %>%
  mutate(poverty_quant = quantcut(inc_below_pov_pct, q = 5, na.rm=TRUE)) %>%
  mutate(black_quant = fct_recode(black_quant, 
           one = levels(black_quant)[1], 
           two = levels(black_quant)[2], 
           three = levels(black_quant)[3], 
           four = levels(black_quant)[4], 
           five = levels(black_quant)[5])) %>%
  mutate(white_quant = fct_recode(white_quant, 
           one = levels(white_quant)[1], 
           two = levels(white_quant)[2], 
           three = levels(white_quant)[3], 
           four = levels(white_quant)[4], 
           five = levels(white_quant)[5])) %>%  
  mutate(asian_quant = fct_recode(asian_quant, 
           one = levels(asian_quant)[1], 
           two = levels(asian_quant)[2], 
           three = levels(asian_quant)[3], 
           four = levels(asian_quant)[4], 
           five = levels(asian_quant)[5])) %>%  
 mutate(native_quant = fct_recode(native_quant, 
           one = levels(native_quant)[1], 
           two = levels(native_quant)[2], 
           three = levels(native_quant)[3], 
           four = levels(native_quant)[4], 
           five = levels(native_quant)[5])) %>%  
  mutate(hispanic_quant = fct_recode(hispanic_quant, 
           one = levels(hispanic_quant)[1], 
           two = levels(hispanic_quant)[2], 
           three = levels(hispanic_quant)[3], 
           four = levels(hispanic_quant)[4], 
           five = levels(hispanic_quant)[5])) %>%
  mutate(poverty_quant = fct_recode(poverty_quant, 
           one = levels(poverty_quant)[1], 
           two = levels(poverty_quant)[2], 
           three = levels(poverty_quant)[3], 
           four = levels(poverty_quant)[4], 
           five = levels(poverty_quant)[5]))


```



```{r,  echo=FALSE, message=FALSE}
#regressions with quants

quant_reg <- feols(data = myWorking, amount ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips  )

quant_reg_yearFE <- feols(data = myWorking, amount ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips + fiscal_year )

quant_reg_log <- feols(data = myWorking, log(amount) ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips  )

quant_reg_yearFE_log <- feols(data = myWorking, log(amount) ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips + fiscal_year )



# race
etable(quant_reg,
       quant_reg_yearFE,
       quant_reg_log,
       quant_reg_yearFE_log,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 



```
`

# New Outcome Variable 
**Next step -- see if these trends show up with a different outcome variable: the number of projects a county gets (aggregate to every year, every 5 years), total amount of funding a county gets (every year, every five years).

## Annual quantity of grants a county receives 

```{r,  echo=FALSE, message=FALSE}

# Number of projects a county gets per year 
myWorking_quantity <- myWorking %>%
  filter(!is.na(county_fips)) %>%
  group_by(fiscal_year, county_fips) %>%
  mutate(annual_quantity_grants = length(amount)) %>%
  mutate(annual_amount = sum(amount)) %>%
  select(-amount, -type ) %>%
  distinct()


# graphs --------------------

# Annual quanitty of grants 
ggplot(myWorking_quantity, aes(x=annual_quantity_grants)) + 
  geom_histogram() +
  theme_minimal() +
  ggtitle("Count of annual projects a county recieves")

ggplot(myWorking_quantity, aes(x=log(annual_quantity_grants))) + 
  geom_histogram() +
  theme_minimal()+
  ggtitle("Count of LOG of annual projects a county recieves")

ggplot(myWorking_quantity, aes(x=fiscal_year)) + 
  geom_histogram(binwidth=1) +
  theme_minimal()+
  ggtitle("Count of annual projects a county recieves by year") +
  scale_x_continuous(breaks=seq(1965,2020,10))


# Annual amount of money a county receives 
ggplot(myWorking_quantity, aes(x=annual_amount)) + 
  geom_histogram() +
  theme_minimal() +
  ggtitle("Count of annual amount of funding a county recieves")

ggplot(myWorking_quantity, aes(x=log(annual_amount))) + 
  geom_histogram() +
  theme_minimal()+
  ggtitle("Count of LOG of annual amount of funding a county recieves")






``` 


### Socioeconomic 

```{r,  echo=FALSE, message=FALSE}

# Initial regressions -------------------


#park type control
type <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~  population| state_fips )

#economic 
poverty <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~ inc_below_pov_pct + population| state_fips )
med_inc <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~ med_income_house + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 
```



### Race

```{r,  echo=FALSE, message=FALSE}

# race
white <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~ white_pct  + population| state_fips )
black <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~ black_pct  + population| state_fips )
native <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~ native_pct  + population| state_fips )
hispanic <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~ hispanic_pct  + population| state_fips )
asian <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~ asian_pct + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# resids <- residuals(one)
```


## Annual amount of funding a county recieves


### Socioeconomic 

```{r,  echo=FALSE, message=FALSE}

# Initial regressions -------------------

#park type control
type <- feols(data = myWorking_quantity, log(annual_amount) ~ population| state_fips )

#economic 
poverty <- feols(data = myWorking_quantity, log(annual_amount) ~ inc_below_pov_pct + population| state_fips )
med_inc <- feols(data = myWorking_quantity, log(annual_amount) ~ med_income_house + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 
```



### Race

```{r,  echo=FALSE, message=FALSE}

# race
white <- feols(data = myWorking_quantity, log(annual_amount) ~ white_pct + population| state_fips )
black <- feols(data = myWorking_quantity, log(annual_amount) ~ black_pct  + population| state_fips )
native <- feols(data = myWorking_quantity, log(annual_amount) ~ native_pct + population| state_fips )
hispanic <- feols(data = myWorking_quantity, log(annual_amount) ~ hispanic_pct + population| state_fips )
asian <- feols(data = myWorking_quantity, log(annual_amount) ~ asian_pct + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# resids <- residuals(one)
```


## Not logged

### Annual quantity of grants a county receives 

#### Socioeconomic 

```{r,  echo=FALSE, message=FALSE}

# Initial regressions -------------------

#park type control
type <- feols(data = myWorking_quantity, annual_quantity_grants ~ population| state_fips )

#economic 
poverty <- feols(data = myWorking_quantity, annual_quantity_grants ~ inc_below_pov_pct  + population| state_fips )
med_inc <- feols(data = myWorking_quantity, annual_quantity_grants ~ med_income_house + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 
```



#### Race

```{r,  echo=FALSE, message=FALSE}

# race
white <- feols(data = myWorking_quantity, annual_quantity_grants ~ white_pct  + population| state_fips )
black <- feols(data = myWorking_quantity, annual_quantity_grants ~ black_pct  + population| state_fips )
native <- feols(data = myWorking_quantity, annual_quantity_grants ~ native_pct  + population| state_fips )
hispanic <- feols(data = myWorking_quantity, annual_quantity_grants ~ hispanic_pct  + population| state_fips )
asian <- feols(data = myWorking_quantity, annual_quantity_grants ~ asian_pct  + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# resids <- residuals(one)
```


### Annual amount of funding a county recieves


#### Socioeconomic 

```{r,  echo=FALSE, message=FALSE}

# Initial regressions -------------------

#park type control
type <- feols(data = myWorking_quantity, annual_amount ~  population| state_fips )

#economic 
poverty <- feols(data = myWorking_quantity, annual_amount ~ inc_below_pov_pct  + population| state_fips )
med_inc <- feols(data = myWorking_quantity, annual_amount ~ med_income_house + population| state_fips )

#economic 
etable(type,
       poverty,
       med_inc,
       order = c("_pct", "med_"),
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Income") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 
```



#### Race

```{r,  echo=FALSE, message=FALSE}

# race
white <- feols(data = myWorking_quantity, annual_amount ~ white_pct + population| state_fips )
black <- feols(data = myWorking_quantity, annual_amount ~ black_pct + population| state_fips )
native <- feols(data = myWorking_quantity, annual_amount ~ native_pct + population| state_fips )
hispanic <- feols(data = myWorking_quantity, annual_amount ~ hispanic_pct + population| state_fips )
asian <- feols(data = myWorking_quantity, annual_amount ~ asian_pct + population| state_fips )



# race
etable(white,
       black,
       asian,
       native,
       hispanic,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 


```

# Quantiles and new outcome variables 

## Annual Quantity 

```{r,  echo=FALSE, message=FALSE}
#regressions with quants

quant_reg <- feols(data = myWorking_quantity, annual_quantity_grants ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips  )

quant_reg_yearFE <- feols(data = myWorking_quantity, annual_quantity_grants ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips + fiscal_year )

quant_reg_log <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips  )

quant_reg_yearFE_log <- feols(data = myWorking_quantity, log(annual_quantity_grants) ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips + fiscal_year )



# race
etable(quant_reg,
       quant_reg_yearFE,
       quant_reg_log,
       quant_reg_yearFE_log,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 



```

## Annual Amount 

```{r,  echo=FALSE, message=FALSE}
#regressions with quants

quant_reg <- feols(data = myWorking_quantity, annual_amount ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips  )

quant_reg_yearFE <- feols(data = myWorking_quantity, annual_amount ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips + fiscal_year )

quant_reg_log <- feols(data = myWorking_quantity, log(annual_amount) ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips  )

quant_reg_yearFE_log <- feols(data = myWorking_quantity, log(annual_amount) ~ white_quant + black_quant + asian_quant + native_quant + hispanic_quant +poverty_quant  + population| state_fips + fiscal_year )



# race
etable(quant_reg,
       quant_reg_yearFE,
       quant_reg_log,
       quant_reg_yearFE_log,
       order = "_pct",
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 



```



Notes to self on next step: may want to make even more race factors (decile). And then make a 0/1 indicator on if a county is that race or not if it is in the top decile. This is a tough analysis to run at the county level (kathrine sims runs this as the tract level).
