==> 04.three_week_update.Rmd <==
---
title: "Three Week Update"
author: "Andie Creel"
date: "6/18/2021"
output: 
  pdf_document:
    number_sections: true
    extra_dependencies: ["float"]

knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, encoding = encoding, output_file = file.path(dirname(inputFile), 'Exploratory_Output/regressions/three_week_update.pdf')) })
---

```{r, include=FALSE}
#R Markdown stuff
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(fig.pos = "!H", out.extra = "")
options(tinytex.verbose = TRUE)
```

```{r setup, include=FALSE}
library(vroom)
library(dplyr)
library(tidyr)
library(stringr)
library(usmap)
library(tidycensus)
library(ggplot2)
library(fixest)
library(kableExtra)
library(gtools)
library(forcats)
library("wesanderson")
options(scipen=99999)
library(janitor)

myWorking <- vroom("Datasets/clean_data/lwcf_tws_projects_onlyACS.csv")

# additional data cleaning ------------

# Removing counties whose FIPSs code I didn't identify 
myWorking <- myWorking %>%
  filter(!is.na(county_fips)) %>%
  distinct()

#removing grants we don't know the size of or that are zero 
myWorking <- myWorking %>%
  filter(!is.na(amount)) %>%
  filter(amount != 0)

myPercentiles <- quantile(myWorking$amount, probs = c(0, 0.05, .1, .25, .5, .75, .9, .95, 1))

#dropping parks in the lowest 5% and in the highest 5% (about the 95th percentile)

myWorking <- myWorking %>%
  filter(amount > myPercentiles["5%"]) %>%
  filter(amount < myPercentiles["95%"])


# Adding metro, nonmetro and rural variable
myUrbanKey <- myWorking %>%
  dplyr::select(RUCC_2013, Description) %>%
  distinct()

myWorking <- myWorking %>%
  mutate(metro = if_else(RUCC_2013 == 1 | RUCC_2013 == 2 | RUCC_2013 == 3, 1, 0)) %>%
  mutate(nonmetro = if_else(RUCC_2013 == 4 | RUCC_2013 == 5 | RUCC_2013 == 6, 1, 0)) %>%
  mutate(rural = if_else(RUCC_2013 == 7 | RUCC_2013 == 8 | RUCC_2013 == 9, 1, 0))


# getting ready for deciles
myWorking <- myWorking %>%
  mutate(type = as.factor(type)) %>%
  mutate(black_dec = quantcut(black_pct, q = 10, na.rm=TRUE)) %>%
  mutate(white_dec = quantcut(white_pct, q = 10, na.rm=TRUE)) %>%
  mutate(asian_dec = quantcut(asian_pct, q = 10, na.rm=TRUE)) %>%
  mutate(native_dec = quantcut(native_pct, q = 10, na.rm=TRUE)) %>%
  mutate(hispanic_dec = quantcut(hispanic_pct, q = 10, na.rm=TRUE)) %>%
  mutate(poverty_dec = quantcut(inc_below_pov_pct, q = 10, na.rm=TRUE)) %>%
  mutate(noCollegeDegree_dec = quantcut(noCollegeDegree_pct, q = 10, na.rm=TRUE)) %>%
  mutate(black_dec = fct_recode(black_dec, 
           ten = levels(black_dec)[10])) %>%
  mutate(white_dec = fct_recode(white_dec, 
           ten = levels(white_dec)[10])) %>%  
  mutate(asian_dec = fct_recode(asian_dec, 
           ten = levels(asian_dec)[10])) %>%  
 mutate(native_dec = fct_recode(native_dec, 
           ten = levels(native_dec)[10])) %>%  
  mutate(hispanic_dec = fct_recode(hispanic_dec, 
           ten = levels(hispanic_dec)[10])) %>%
  mutate(poverty_dec = fct_recode(poverty_dec, 
           ten = levels(poverty_dec)[10])) %>%
  mutate(noCollegeDegree_dec = fct_recode(noCollegeDegree_dec, 
           ten = levels(noCollegeDegree_dec)[10]))

# making 0/1 for counties in the highest decile for different races and ethnicities 
myWorking <- myWorking %>%
  mutate(white_county = if_else(white_dec == "ten", 1, 0)) %>%
  mutate(black_county = if_else(black_dec == "ten", 1, 0)) %>%
  mutate(asian_county = if_else(asian_dec == "ten", 1, 0)) %>%
  mutate(native_county = if_else(native_dec == "ten", 1, 0)) %>%
  mutate(hispanic_county = if_else(hispanic_dec == "ten", 1, 0)) %>%
  mutate(noCollegeDegree_county = if_else(hispanic_dec == "ten", 1, 0)) 

# Removing useless variables 
myWorking <- myWorking %>%
  dplyr::select(-c(objectid, county, grantidelement, grantelementtitle, grantsponsor, congdistrict, relate))

# other outcome variable -------------------------------------
# annual quantity of grants

 myCounts <- myWorking %>% 
  group_by(fiscal_year, county_fips) %>%
  tally() %>%
  dplyr::rename(annual_quantity = n)
 
 myWorking_quantity <- left_join(myWorking, myCounts , by = c("fiscal_year", "county_fips")) %>%
   dplyr::select(-type, -amount) %>%
   distinct()

# annual amount of funding
mySum <- myWorking %>%
   group_by(fiscal_year, county_fips) %>%
    dplyr::summarise(annual_amount = sum(amount))

  
myWorking_quantity <- left_join(myWorking_quantity, mySum , by = c("fiscal_year", "county_fips"))



# Making percentages to be whole numbers 
myWorking <- myWorking %>%
  mutate(white_pct = white_pct *100) %>%
  mutate(black_pct = black_pct *100) %>%
  mutate(asian_pct = asian_pct *100) %>%
  mutate(native_pct = native_pct *100) %>%
  mutate(hispanic_pct = hispanic_pct *100) %>%
  mutate(noCollegeDegree_pct = noCollegeDegree_pct *100) %>%
  mutate(inc_below_pov_pct = inc_below_pov_pct *100) 

# Function to compute regression result 


myLogLin <- function(x){
  y <- (exp(x)-1)*100 
  return(y)
}


```




# Percent Race/Ethnicity Measures


```{r,  echo=FALSE, message=FALSE, fig.width=3.5, fig.height=2.5}
#regressing all races together

# Median income 
one <- feols(data = myWorking, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + type + population | state_fips + fiscal_year)
two <- feols(data = myWorking, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + type + population| state_fips + fiscal_year)
three <- feols(data = myWorking, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type + population| state_fips + fiscal_year)
four <- feols(data = myWorking, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type  + rural| state_fips + fiscal_year)

etable(one,
       two,
       three,
       four,
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Percent Race/Ethnicity Measures") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# qqnorm(residuals(one))
# qqnorm(residuals(two))
# qqnorm(residuals(three))
# qqnorm(residuals(four))

```

interpretation of units for continuous variables in log-linear regression: The effect for a 1 unit increase in x results in a $(e^\beta -1)*100$ change in y.

interpretation of units for dummy variables in log-linear regression: The effect of "turning on" the dummy variable x results in a $(e^\beta -1)*100$ change in y.


Observations:

* a 1 % increase in the percent of the county that's **white** results in a 0.6% **decrease** in the size of grant award
* a 1% increase in the percent of the county that's **hispanic** results in a 0.6% **increase** in the size of the grant award
* a 1% increase in the percent of the county that **lacks a college degree** results in a 1.9% **decrease** in the size of the grant award
* "turning on" **rural** results in a 9.8% **decrease** in the size of the grant award


# Decile Regressions
Measuring race/ethnicity as a 0/1 indicator for whether or not the county is in the top decile for % of total population that is that is that race/ethnicity

```{r,  echo=FALSE, message=FALSE, fig.width=3.5, fig.height=2.5}
#regressing all races together

# Median income 
one <- feols(data = myWorking, log(amount) ~ white_county + black_county + asian_county + hispanic_county +native_county + type + population | state_fips + fiscal_year)
two <- feols(data = myWorking, log(amount) ~ white_county + black_county + asian_county + hispanic_county +native_county + med_income_house + type + population| state_fips + fiscal_year)
three <- feols(data = myWorking, log(amount) ~ white_county + black_county + asian_county + hispanic_county +native_county + med_income_house + noCollegeDegree_county + type + population| state_fips + fiscal_year)
four <- feols(data = myWorking, log(amount) ~ white_county + black_county + asian_county + hispanic_county +native_county + med_income_house + noCollegeDegree_county + type  + rural| state_fips + fiscal_year)

etable(one,
       two,
       three,
       four,
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Decile Regressions") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# qqnorm(residuals(one))
# qqnorm(residuals(two))
# qqnorm(residuals(three))
# qqnorm(residuals(four))

```

*No College degree and rural are too correlated to regress together (and maybe hispanic) when using deciles*

* Turning on "**white** county" (*i.e.* being in the top 10% of counties with the largest amount of people identifying as white) results in a **11% decrease** in grant award sizes 
* Turning on "**black** county" (*i.e.* being in the top 10% of counties with the largest amount of people identifying as black) results in a **25% increase** in grant award sizes 
* Turning on "**asian** county" (*i.e.* being in the top 10% of counties with the largest amount of people identifying as asian) results in a **26% increase** in grant award sizes
* Turning on "**hispanic** county" (*i.e.* being in the top 10% of counties with the largest amount of people identifying as hispanic) results in a **14% increase** in grant award sizes
*  Turning on **rural** results in a **14% decrease** in grant award sizes


# Results for Development Grants
Grouping development and redevelopment grants together and only looking at those 

```{r,  echo=FALSE, message=FALSE, fig.width=3.5, fig.height=2.5}
#getting ready to filter on development 
myWorking <- myWorking %>%
  mutate(type = as.character(type)) %>%
  mutate(type = str_sub(type, 1, 1))

#only looking at development and redevelopment results
myWorking_dev <- myWorking %>%
  dplyr::filter(type == "D" | type == "R")


#regressing all races together

# Median income 
one <- feols(data = myWorking_dev, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + type + population | state_fips + fiscal_year)
two <- feols(data = myWorking_dev, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + type + population| state_fips + fiscal_year)
three <- feols(data = myWorking_dev, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type + population| state_fips + fiscal_year)
four <- feols(data = myWorking_dev, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type  + rural| state_fips + fiscal_year)

etable(one,
       two,
       three,
       four,
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Results for Development Grants") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# qqnorm(residuals(one))
# qqnorm(residuals(two))
# qqnorm(residuals(three))
# qqnorm(residuals(four))

```


Observations:

* a 1 % increase in the percent of the county that's **white** results in a 0.6% **decrease** in the size of grant award
* a 1% increase in the percent of the county that's **asian** results in a 1% **increase** in the size of the grant award
* a 1% increase in the percent of the county that's **hispanic** results in a 0.5% **increase** in the size of the grant award
* a 1% increase in the percent of the county that **lacks a college degree** results in a 1.7% **decrease** in the size of the grant award
* turning on **rural** results in a 12% **decrease** in the size of the grant award

These are essentially the same as when looking at all grant types, but asian is more precisely estimated. 


# Per Capita Outcomes
Grant amount divided by total population


```{r,  echo=FALSE, message=FALSE, fig.width=3.5, fig.height=2.5}
#calculating per capita outcome variable 
myWorking_perCap <- myWorking %>%
  dplyr::mutate(amount = amount / population)

#regressing all races together

# Median income 
one <- feols(data = myWorking_perCap, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + type  | state_fips + fiscal_year)
two <- feols(data = myWorking_perCap, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + type | state_fips + fiscal_year)
three <- feols(data = myWorking_perCap, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type | state_fips + fiscal_year)
four <- feols(data = myWorking_perCap, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type  + rural| state_fips + fiscal_year)

etable(one,
       two,
       three,
       four,
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Per Capita Outcomes") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# qqnorm(residuals(one))
# qqnorm(residuals(two))
# qqnorm(residuals(three))
# qqnorm(residuals(four))

```


**Note: when I use a per capita measure, results flip**

* a 1 % increase in the percent of the county that's **hispanic** results in a 2.5% **decrease** in the *award size per capita*
* a 1 % increase in the percent of the county that **lacks a college degree** results in a 8% **increase** in the *award size per capita*
* Turning on **rural** results in a 194% **increase** in the *award size per capita*

# Per capita results for Development Grants

```{r,  echo=FALSE, message=FALSE, fig.width=3.5, fig.height=2.5}
#getting ready to filter on development 
myWorking <- myWorking %>%
  mutate(type = as.character(type)) %>%
  mutate(type = str_sub(type, 1, 1))

#only looking at PER CAPITA development and redevelopment results
myWorking_dev <- myWorking %>%
  dplyr::filter(type == "D" | type == "R") %>%
  dplyr::mutate(amount = amount / population)


#regressing all races together

# Median income 
one <- feols(data = myWorking_dev, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + type  | state_fips + fiscal_year)
two <- feols(data = myWorking_dev, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + type | state_fips + fiscal_year)
three <- feols(data = myWorking_dev, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type | state_fips + fiscal_year)
four <- feols(data = myWorking_dev, log(amount) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type  + rural| state_fips + fiscal_year)

etable(one,
       two,
       three,
       four,
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Per capita results for Development Grants") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# qqnorm(residuals(one))
# qqnorm(residuals(two))
# qqnorm(residuals(three))
# qqnorm(residuals(four))

```

**Note: when I use a per capita measure for development grants, results flip**



* a 1 % increase in the percent of the county that's **hispanic** results in a 2.4% **decrease** in the *award size per capita*
* a 1 % increase in the percent of the county that **lacks a college degree** results in a 8% **increase** in the *award size per capita*
* Turning on **rural** results in a 191% **increase** in the *award size per capita*


# Percent of total


```{r,  echo=FALSE, message=FALSE}


myWhole <- myWorking %>%
  dplyr::group_by(fiscal_year) %>%
  dplyr::summarise(year_total_amount = sum(amount))

myWorking <- left_join(myWorking, myWhole, by = "fiscal_year")

myWorking <- myWorking %>%
  dplyr::mutate(amount_percent = (amount / year_total_amount)*100)

#regressing all races together

# Median income 
one <- feols(data = myWorking, log(amount_percent) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + type + population | state_fips + fiscal_year)
two <- feols(data = myWorking, log(amount_percent) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + type + population| state_fips + fiscal_year)
three <- feols(data = myWorking, log(amount_percent) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type + population| state_fips + fiscal_year)
four <- feols(data = myWorking, log(amount_percent) ~ white_pct + black_pct + asian_pct + hispanic_pct +native_pct + med_income_house + noCollegeDegree_pct + type  + rural| state_fips + fiscal_year)

etable(one,
       two,
       three,
       four,
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Percent of total") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# qqnorm(residuals(one))
# qqnorm(residuals(two))
# qqnorm(residuals(three))
# qqnorm(residuals(four))

```

*These regression results are the exact same as the "1. Percent Race/Ethnicity Measures" because I calculated the % of total with a linear-monotonic transformation (divided it all by the same total and multipled it all by 100). The R2 is different because there is less variation with this transformation*


Observations:

* a 1 % increase in the percent of the county that's **white** results in a 0.6% **decrease** in the size of grant award
* a 1% increase in the percent of the county that's **hispanic** results in a 0.6% **increase** in the size of the grant award
* a 1% increase in the percent of the county that **lacks a college degree** results in a 1.9% **decrease** in the size of the grant award
* "turning on" **rural** results in a 9.8% **decrease** in the size of the grant award
