==> 05.descriptive_Poverty.Rmd <==
---
title: "Poverty Descriptive Maps"
author: "Andie Creel"
date: "7/7/2021"
output: 
  pdf_document:
    number_sections: true
    extra_dependencies: ["float"]

knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, encoding = encoding, output_file = file.path(dirname(inputFile), 'Exploratory_Output/descriptive/poverty.pdf')) })
---

Brown colors tend to indicated positive numbers and purple tend to indicate negative numbers. 

```{r, include=FALSE}
#R Markdown stuff
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(fig.pos = "!H", out.extra = "")
options(tinytex.verbose = TRUE)
```

```{r setup, include=FALSE}
library(vroom)
library(tidyr)
library(stringr)
library(usmap)
library(tidycensus)
library(ggplot2)
library(fixest)
library(kableExtra)
library(gtools)
library(forcats)
library("wesanderson")
options(scipen=99999)
library(janitor)
library(dplyr)
# ---- #
library(choroplethr)
library(choroplethrMaps)
library(usdata)
library(maps)
data(fips_codes)
data(continental_us_states)
library(R6)
# --- #



myWorking <- vroom("Datasets/clean_data/lwcf_annualPop_deccenialDemographics.csv")

# additional data cleaning ------------

# Removing counties whose FIPSs code I didn't identify 
myWorking <- myWorking %>%
  filter(!is.na(fips)) %>%
  distinct() %>%
  mutate(type = as.factor(type)) %>%
  mutate(state_fips = as.factor(str_sub(fips, 1, 2))) %>%
  filter(!is.na(white_pct)) %>%
  dplyr::mutate(POC_pct = 100 - white_pct)

# Getting quartiles (Sims 2020)

#Getting the quartile for each state each year! 
myQuants <- myWorking %>%
  dplyr::group_by(state_fips, year) %>%
  dplyr::summarise(quarts = list(quantile(inc_below_pov_pct, probs = c(.25, .5, .75)))) %>%
  unnest_wider(quarts) %>%
  dplyr::rename(pov_25 = `25%`) %>%
  dplyr::rename(pov_50 = `50%`) %>%
  dplyr::rename(pov_75 = `75%`)

#joining back in
myWorking <- left_join(myWorking, myQuants, by = c("state_fips", "year"))
rm(myQuants)

#making a factor variable for the quartiles (by state and year)
myWorking <- myWorking %>%
  mutate(pov_quarts = if_else(inc_below_pov_pct > pov_75, "4", "")) %>% # fourth quartile 
  mutate(pov_quarts = if_else(inc_below_pov_pct <= pov_75, "3", pov_quarts)) %>% # third quartile
  mutate(pov_quarts = if_else(inc_below_pov_pct <= pov_50, "2", pov_quarts)) %>% # second quartile
  mutate(pov_quarts = if_else(inc_below_pov_pct <= pov_25, "1", pov_quarts)) %>% # first quartile 
  select(-pov_75, -pov_50, - pov_25) %>%
  mutate(pov_quarts = as.factor(pov_quarts))

```


```{r DATA-MANIPULATION, echo=FALSE, message=FALSE}

# --- data manipulation to get quantity of grants per quartile--- #

# total grants per state 
myTotal <- myWorking %>%
  group_by(state_fips) %>%
  mutate(state_total = sum(got_grant)) %>%
  select(state_fips, state_total) %>%
  distinct()

# number of grants given to first quart counties ever per state
myOne <- myWorking %>%
  filter(pov_quarts == "1") %>%
  group_by(state_fips) %>%
  mutate(one_total = sum(got_grant)) %>%
  select(state_fips, one_total) %>%
  distinct()

# number of grants given to second quart counties ever per state
myTwo <- myWorking %>%
  filter(pov_quarts == "2") %>%
  group_by(state_fips) %>%
  mutate(two_total = sum(got_grant)) %>%
  select(state_fips, two_total) %>%
  distinct()

# number of grants given to third quart counties ever per state
myThree <- myWorking %>%
  filter(pov_quarts == "3") %>%
  group_by(state_fips) %>%
  mutate(three_total = sum(got_grant)) %>%
  select(state_fips, three_total) %>%
  distinct()

# number of grants given to third quart counties ever per state
myFour <- myWorking %>%
  filter(pov_quarts == "4") %>%
  group_by(state_fips) %>%
  mutate(four_total = sum(got_grant)) %>%
  select(state_fips, four_total) %>%
  distinct()

#merge
myPov_mapData <- left_join(myTotal, myOne , by = "state_fips")
myPov_mapData <- left_join(myPov_mapData, myTwo , by = "state_fips")
myPov_mapData <- left_join(myPov_mapData, myThree , by = "state_fips")
myPov_mapData <- left_join(myPov_mapData, myFour , by = "state_fips")
rm(myTotal, myOne, myTwo, myThree, myFour)

#percents 
myPov_mapData <- myPov_mapData %>%
  mutate(one_pct = round(one_total/state_total * 100, 0 )) %>%
  mutate(two_pct = round(two_total/state_total * 100, 0 )) %>%
  mutate(three_pct = round(three_total/state_total * 100 , 0)) %>%
  mutate(four_pct = round(four_total/state_total * 100, 0 )) 
```


**Poverty Quartiles**

I calculate the percentage of a county's population that is below the poverty line using data from the decennial census. 

I then categorize counties into four quartiles based on the % of the population that is in poverty The first quartile categorizes the least impoverished counties and the fourth quartile categorizes the counties with the most people under the poverty line. 

When calculating quartiles, I group by state and decade. I group by state because the LWCF is administered at the state level, therefore managers are making **intrastate** comparisons when deciding how to allocate funds. I group by decade because the % of the population that is in poverty changes through time. Therefore, a county that had a significant number of people in poverty in the 1970s may have experienced socioeconomic changes by the 2010s. I group at the decade level because my demographic characteristics come from the decennial census which is collected at the decade level.  


# Percent of Grants Given to Each Quartile 

For each of the four quartiles, I calculate the percentage received of total grants given from 1966 until 2018. This metric is a helpful moment for the total quantity of grants given to the different quartiles, but ignores time trends as well as the amount of funding allocated. 

## Maps

```{r, echo=FALSE, message=FALSE, warning=FALSE}

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myPov_mapData <- left_join(myPov_mapData, fips_codes, by = c("state_fips" = "state_code")) 

myPov_mapData <- myPov_mapData %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region))

# --- maps --- #

# First quartile
myOne_map <- myPov_mapData %>%
  select(region, one_pct) %>%
  rename(value = one_pct) %>%
  filter(!is.na(region))

state_choropleth(myOne_map,
  title = "% of grants given to first quartile of poverty counties",
  legend = "Percent %",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr")+
  ggtitle(label = "% of total grants given to first quartile poverty counties", subtitle =  "(richest counties)") + 
  labs(fill = "% received by 1st")

# Second quartile
myTwo_map <- myPov_mapData %>%
  select(region, two_pct) %>%
  rename(value = two_pct) %>%
  filter(!is.na(region))

state_choropleth(myTwo_map,
  title = "% of grants given to second quartile of poverty counties",
  legend = "Percent %",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "% received by 2nd")


# Third quartile
myThree_map <- myPov_mapData %>%
  select(region, three_pct) %>%
  rename(value = three_pct) %>%
  filter(!is.na(region))

state_choropleth(myThree_map,
  title = "% of grants given to third quartile of poverty counties",
  legend = "Percent %",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr")+ 
  labs(fill = "% received by 3rd")

# Four quartile
myFour_map <- myPov_mapData %>%
  select(region, four_pct) %>%
  rename(value = four_pct) %>%
  filter(!is.na(region))

state_choropleth(myFour_map,
  title = "% of grants given to fourth quartile of poverty counties",
  legend = "Percent %",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "% received by 4th") + 
    ggtitle(label = "% of grants given to fourth quartile of poverty counties", subtitle =  "(counties with most poverty)") 

# difference between fourth and first quartile 

myDifference_map <- myPov_mapData %>%
  select(region, one_pct, four_pct) %>%
  mutate(value = four_pct - one_pct) %>%
  filter(!is.na(region))

state_choropleth(myDifference_map,
  title = "Difference in % grants recieved by fourth quartile and first",
  legend = "Difference in % recieved",
  zoom = continental_us_states,
  num_colors = 5) +
  scale_fill_brewer(palette="Purples", direction = -1) + 
  labs(fill = "Difference in % recieved")


```

**Interpretation:** 

* In Montana, the counties in the first quartile for percentage of population that is in poverty (ie the richest counties) receive 37-39% of the quantity of grants given to Montana. 
* In Montana, the counties in the fourth quartile for % of population that is a in poverty (counties with the most poverty) receive 14-16% of the quantity of grants given to Montana. 
* In Montana, counties in the fourth quartile receive 23-20% more grants than counties in the first quartile (calculated using a simple diference). 

## Table 

```{r, echo=FALSE, message=FALSE, warning=FALSE}

#get percent of Poverty for the state

state_pct_Poverty <- myWorking %>%
  select(state_fips, inc_below_pov_pct) %>%
  distinct() %>%
  group_by(state_fips) %>%
  mutate(avg_inc_below_pov_pct = mean(inc_below_pov_pct)) %>%
  select(state_fips, avg_inc_below_pov_pct) %>%
  distinct()

myPoverty_table <- left_join(myPov_mapData, state_pct_Poverty, by = "state_fips" )

myPoverty_table <- myPoverty_table %>%
  ungroup() %>%
  select(region, one_pct, two_pct, three_pct, four_pct, avg_inc_below_pov_pct) %>%
  mutate(difference = four_pct - one_pct) %>%
  filter(region != "district of columbia") %>%
  filter(region != "puerto rico") %>%
  arrange(-avg_inc_below_pov_pct) %>%
  mutate(avg_inc_below_pov_pct = round(avg_inc_below_pov_pct, digits = 2)) %>%
  mutate(region = str_to_title(region)) %>%
  select(region, one_pct, two_pct, three_pct, four_pct, difference, avg_inc_below_pov_pct)


myPoverty_table %>%
  kable(booktabs = T, caption = "% grants recieved by quartiles ",
      col.names = c("State",
                    "First",
                    "Second", 
                    "Third",
                    "Fourth",
                    "Difference (four minus first)", 
                    "Avg % of county that's in poverty")) %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

# Plot of avg % Poverty in state and difference between quarts 
ggplot(myPoverty_table, aes(x=avg_inc_below_pov_pct, y=difference)) + geom_point()+ geom_smooth(method = "lm") +theme_bw() + 
  xlab("Avg. % of in poverty per County") +
  ylab("Difference in fouth and first quartile") + ggtitle("% grants recieved")

```
When looking at the difference in the % of grants received by quartile, there is a slight downward trend in the fourth quartile receiving less than the first as a state's avg % of counties' population in poverty increases.

This may indicate that as the % of a counties that is in poverty is more observable (because there is more poverty to notice) the disparity in the distribution of funds gets more severe. 

## Time Differences: % Recieved 

```{r, echo=FALSE, message=FALSE, warning=FALSE}

myTime_pct <- function(myStart = 1965, myEnd = 2018){
  
  myWorking <- myWorking %>%
    dplyr::filter(year >= myStart) %>%
    dplyr::filter(year < myEnd)

  
  
# Removing counties whose FIPSs code I didn't identify 
myWorking <- myWorking %>%
  filter(!is.na(fips)) %>%
  distinct() %>%
  mutate(type = as.factor(type)) %>%
  mutate(state_fips = as.factor(str_sub(fips, 1, 2))) %>%
  filter(!is.na(white_pct)) %>%
  dplyr::mutate(POC_pct = 100 - white_pct)

# Getting quartiles (Sims 2020)

#Getting the quartile for each state each year! 
myQuants <- myWorking %>%
  dplyr::group_by(state_fips, year) %>%
  dplyr::summarise(quarts = list(quantile(inc_below_pov_pct, probs = c(.25, .5, .75)))) %>%
  unnest_wider(quarts) %>%
  dplyr::rename(pov_25 = `25%`) %>%
  dplyr::rename(pov_50 = `50%`) %>%
  dplyr::rename(pov_75 = `75%`)

#joining back in
myWorking <- left_join(myWorking, myQuants, by = c("state_fips", "year"))
rm(myQuants)

#making a factor variable for the quartiles (by state and year)
myWorking <- myWorking %>%
  mutate(pov_quarts = if_else(inc_below_pov_pct > pov_75, "4", "")) %>% # fourth quartile 
  mutate(pov_quarts = if_else(inc_below_pov_pct <= pov_75, "3", pov_quarts)) %>% # third quartile
  mutate(pov_quarts = if_else(inc_below_pov_pct <= pov_50, "2", pov_quarts)) %>% # second quartile
  mutate(pov_quarts = if_else(inc_below_pov_pct <= pov_25, "1", pov_quarts)) %>% # first quartile 
  select(-pov_75, -pov_50, - pov_25) %>%
  mutate(pov_quarts = as.factor(pov_quarts))

# --- data manipulation to get quantity of grants per quartile--- #

# total grants per state 
myTotal <- myWorking %>%
  group_by(state_fips) %>%
  mutate(state_total = sum(got_grant)) %>%
  select(state_fips, state_total) %>%
  distinct()

# number of grants given to first quart counties ever per state
myOne <- myWorking %>%
  filter(pov_quarts == "1") %>%
  group_by(state_fips) %>%
  mutate(one_total = sum(got_grant)) %>%
  select(state_fips, one_total) %>%
  distinct()

# number of grants given to second quart counties ever per state
myTwo <- myWorking %>%
  filter(pov_quarts == "2") %>%
  group_by(state_fips) %>%
  mutate(two_total = sum(got_grant)) %>%
  select(state_fips, two_total) %>%
  distinct()

# number of grants given to third quart counties ever per state
myThree <- myWorking %>%
  filter(pov_quarts == "3") %>%
  group_by(state_fips) %>%
  mutate(three_total = sum(got_grant)) %>%
  select(state_fips, three_total) %>%
  distinct()

# number of grants given to third quart counties ever per state
myFour <- myWorking %>%
  filter(pov_quarts == "4") %>%
  group_by(state_fips) %>%
  mutate(four_total = sum(got_grant)) %>%
  select(state_fips, four_total) %>%
  distinct()

#merge
myPov_mapData <- left_join(myTotal, myOne , by = "state_fips")
myPov_mapData <- left_join(myPov_mapData, myTwo , by = "state_fips")
myPov_mapData <- left_join(myPov_mapData, myThree , by = "state_fips")
myPov_mapData <- left_join(myPov_mapData, myFour , by = "state_fips")
rm(myTotal, myOne, myTwo, myThree, myFour)

#percents 
myPov_mapData <- myPov_mapData %>%
  mutate(one_pct = round(one_total/state_total * 100, 0 )) %>%
  mutate(two_pct = round(two_total/state_total * 100, 0 )) %>%
  mutate(three_pct = round(three_total/state_total * 100 , 0)) %>%
  mutate(four_pct = round(four_total/state_total * 100, 0 )) 

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myPov_mapData <- left_join(myPov_mapData, fips_codes, by = c("state_fips" = "state_code")) 

myPov_mapData <- myPov_mapData %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region))

# --- maps --- #

myDifference_map <- myPov_mapData %>%
  select(region, one_pct, four_pct) %>%
  mutate(value = four_pct - one_pct) %>%
  filter(!is.na(region))

state_choropleth(myDifference_map,
  title = paste("Difference in % grants recieved by fourth quartile and first:", myStart, "-", myEnd),
  legend = "Difference in % recieved",
  zoom = continental_us_states,
  num_colors = 5) +
  scale_fill_brewer(palette="Purples", direction = -1) + 
  labs(fill = "Difference in % recieved")

}

myTime_pct()
myTime_pct(myEnd = 1981)
myTime_pct(myStart = 1981, myEnd = 2008)
myTime_pct(myStart = 2008)




```
The pattern of counties in the quartile with the most people in poverty receiving significantly less grants than the white counties varries quite a bit through time, I can't see an obvious trend for any region so far! 


# Average Grant Size per Capita

I calculate the per capita amount of grant funding for each year and include years when the amount is zero. I find the average per capita amount of grant funding for each quartile. 

## Maps

```{r DATA-MANIPULATION-2, echo=FALSE, message=FALSE, warning= FALSE}

# --- data manipulation to get amount per person --- #

#getting region in here for the state_choropleth function
myWorking <- left_join(myWorking, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region))

# calculating grant award per capita for each county each year
myWorking <- myWorking %>%
  mutate(amount_perCap = if_else(!is.na(amount), amount / population,  0))

# -- first quartile maps -- #

# First quartile 
myPerCap_map1 <- myWorking %>%
  filter(pov_quarts == "1") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)
  

state_choropleth(myPerCap_map1,
  title = "Average per capita grants amount given to first quartile of poverty counties",
  legend = "Average per cap. $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr")+
  ggtitle(label = "Average per capita grants amount given to first quartile of poverty counties", 
          subtitle =  "(Richest counties)") + 
  labs(fill = "Average per cap. $")

#second quartile 
myPerCap_map2 <- myWorking %>%
  filter(pov_quarts == "2") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)
  

state_choropleth(myPerCap_map2,
  title = "Average per capita grants amount given to second quartile of poverty counties",
  legend = "Average per cap. $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "Average per cap. $")

  
#third quartile 
myPerCap_map3 <- myWorking %>%
  filter(pov_quarts == "3") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)
  

state_choropleth(myPerCap_map3,
  title = "Average per capita grants amount given to thrid quartile of poverty counties",
  legend = "Average per cap. $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "Average per cap. $")

#fourth quartile 
myPerCap_map4 <- myWorking %>%
  filter(pov_quarts == "4") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)
  
state_choropleth(myPerCap_map4,
  title = "Average per capita grants amount given to fourth quartile of poverty counties",
  legend = "Average per cap. $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "Average per cap. $")


# -- difference -- #


myPerCap_difference_map <- left_join(x = myPerCap_map4, y = myPerCap_map1, by = "region") %>%
  rename(value_4 = value.x) %>%
  rename(value_1 = value.y) %>%
  mutate(value = value_4 - value_1) %>%
  filter(!is.na(region))

state_choropleth(myPerCap_difference_map,
  title = "Difference in avg. per capita grant between fourth quartile and first",
  legend = "Difference in $",
  zoom = continental_us_states,
  num_colors = 7) +
  scale_fill_brewer(palette="PuOr", direction = -1) + 
  labs(fill = "Difference in $")




```


**Interpretation:** 

* On average in California, the counties in the first quartile for percentage of population that is in poverty (ie the richest counties) receive \$ 0.30 to \$0.36 of grant funding per person. 
* On average in California, the counties in the fourth quartile for % of population that is in poverty (counties with the most poverty) receive \$ 0.00 to \$0.21 of grant funding per person. 
* In California, counties in the fourth quartile receive \$ -0.10 - \$ 0.01 in difference of grant funding per person than counties in the first quartile. 

## Table

```{r, echo=FALSE, message=FALSE, warning=FALSE}

# Getting columns together 
myPerCap_table <- left_join(myPerCap_map1, myPerCap_map2, by = "region") %>%
  rename(perCap_1 = value.x) %>%
  rename(perCap_2 = value.y) 
myPerCap_table <- left_join(myPerCap_table, myPerCap_map3, by = "region") %>%
  rename(perCap_3 = value) 
myPerCap_table <- left_join(myPerCap_table, myPerCap_map4, by = "region") %>%
  rename(perCap_4 = value) 
myPerCap_table <- left_join(myPerCap_table, myPerCap_difference_map, by = "region") %>%
  rename(perCap_diff = value) %>%
  select(-value_4, -value_1)

#get percent of poverty for the state
state_pct_poverty <- myWorking %>%
  select(region, inc_below_pov_pct) %>%
  distinct() %>%
  group_by(region) %>%
  mutate(avg_inc_below_pov_pct = mean(inc_below_pov_pct)) %>%
  select(region, avg_inc_below_pov_pct) %>%
  distinct()

#rounding two digets for all columns
myPerCap_table <- left_join(myPerCap_table, state_pct_poverty, by = "region") %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  filter(region != "district of columbia") %>%
  filter(region != "puerto rico") 

myPerCap_table %>%
  arrange(-avg_inc_below_pov_pct) %>%
  mutate(region = str_to_title(region)) %>%
  kbl(booktabs = T, 
      caption = "Avg. per capita grant ($)",
      col.names = c("State",
                    "First",
                    "Second", 
                    "Third",
                    "Fourth",
                    "Difference (four minus first)", 
                    "Avg % of county that's in poverty")) %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 


# Plot of avg % poverty in state and difference between quarts 
ggplot(myPerCap_table, aes(x=avg_inc_below_pov_pct, y=perCap_diff)) + geom_point() + geom_smooth(method = "lm") +theme_bw() + 
  xlab("Avg. % of people in poverty per County") +
  ylab("Difference in fouth and first quartile") + ggtitle("Avg. per capita spending")

```

When looking at the difference in the avg. per capita funding received, there is not an obvious trend with a counties' population that is in poverty. This may be due to the amount of funding a state receives being highly correlated with population and so this final figure is not helpful for a per capita measure because grant funding is allocated to states in a way that tries to guarantee that per capita funding is equal across states.


## Time Differences: Per captia grant award ($)

```{r, echo=FALSE, message=FALSE, warning=FALSE}

myTime_perCap <- function(myStart = 1965, myEnd = 2018){
myWorking <- myWorking %>%
    dplyr::filter(year >= myStart) %>%
    dplyr::filter(year < myEnd)
  
# First quartile 
myPerCap_map1 <- myWorking %>%
  filter(pov_quarts == "1") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)

# Fourth quartile 
myPerCap_map4 <- myWorking %>%
  filter(pov_quarts == "4") %>%
  group_by(region) %>%
  mutate(mean_amount_perCap = mean(amount_perCap)) %>%
  select(region, mean_amount_perCap) %>%
  distinct() %>%
  rename(value = mean_amount_perCap)

# -- difference -- #
myPerCap_difference_map <- left_join(x = myPerCap_map4, y = myPerCap_map1, by = "region") %>%
  rename(value_4 = value.x) %>%
  rename(value_1 = value.y) %>%
  mutate(value = value_4 - value_1) %>%
  filter(!is.na(region))


state_choropleth(myPerCap_difference_map,
  title = paste("Difference in avg. per capita grant between fourth quartile and first", myStart, "-", myEnd),
  legend = "Difference in $",
  zoom = continental_us_states,
  num_colors = 7) +
  scale_fill_brewer(palette="PuOr", direction = -1) + 
  labs(fill = "Difference in $")


}

myTime_perCap()
myTime_perCap(myEnd = 1981)
myTime_perCap(myStart = 1981, myEnd = 2008)
myTime_perCap(myStart = 2008)


```

I am not observing an obvious time trend for any specific regions thus far... California has a really different trend post-Reagan years and post Obama years. 

** I'm trying think of way to organize this so that I more easily observe time trends then only looking at the maps**


# Average Grant Size (Total)

I find the average amount of grant funding for each quartile and include years when the amount is zero.

## Maps

```{r DATA-MANIPULATION-3, echo=FALSE, message=FALSE, warning= FALSE}

# --- data manipulation to get amount per person --- #

# calculating grant award for each county each year
myWorking <- myWorking %>%
  mutate(amount= if_else(!is.na(amount), amount,  0))

# -- first quartile maps -- #

# First quartile 
myAmount_map1 <- myWorking %>%
  filter(pov_quarts == "1") %>%
  group_by(region) %>%
  mutate(mean_amount = mean(amount)) %>%
  select(region, mean_amount) %>%
  distinct() %>%
  rename(value = mean_amount)
  

state_choropleth(myAmount_map1,
  title = "Average grants amount given to first quartile of poverty counties",
  legend = "Average $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr")+
  ggtitle(label = "Average grants amount given to first quartile of poverty counties", 
          subtitle =  "(Richest counties)") + 
  labs(fill = "Average $")

#second quartile 
myAmount_map2 <- myWorking %>%
  filter(pov_quarts == "2") %>%
  group_by(region) %>%
  mutate(mean_amount = mean(amount)) %>%
  select(region, mean_amount) %>%
  distinct() %>%
  rename(value = mean_amount)
  

state_choropleth(myAmount_map2,
  title = "Average grants amount given to second quartile of poverty counties",
  legend = "Average $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "Average $")

  
#third quartile 
myAmount_map3 <- myWorking %>%
  filter(pov_quarts == "3") %>%
  group_by(region) %>%
  mutate(mean_amount = mean(amount)) %>%
  select(region, mean_amount) %>%
  distinct() %>%
  rename(value = mean_amount)
  

state_choropleth(myAmount_map3,
  title = "Average grants amount given to thrid quartile of poverty counties",
  legend = "Average $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "Average $")

#fourth quartile 
myAmount_map4 <- myWorking %>%
  filter(pov_quarts == "4") %>%
  group_by(region) %>%
  mutate(mean_amount = mean(amount)) %>%
  select(region, mean_amount) %>%
  distinct() %>%
  rename(value = mean_amount)
  
state_choropleth(myAmount_map4,
  title = "Average grants amount given to fourth quartile of poverty counties",
  legend = "Average $",
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr") + 
  labs(fill = "Average $")


# -- difference -- #


myAmount_difference_map <- left_join(x = myAmount_map4, y = myAmount_map1, by = "region") %>%
  rename(value_4 = value.x) %>%
  rename(value_1 = value.y) %>%
  mutate(value = value_4 - value_1) %>%
  filter(!is.na(region))

state_choropleth(myAmount_difference_map,
  title = "Difference in avg. grant between fourth quartile and first",
  legend = "Difference in $",
  zoom = continental_us_states,
  num_colors = 5) +
  scale_fill_brewer(palette="Purples", direction = -1) + 
  labs(fill = "Difference in $")




```


**Interpretation:** 

* On average in Nevada, the counties in the first quartile for percentage of population that is in poverty (ie the richest counties) receive \$ 35K to \$74K of grant funding per person. 
* On average in Nevada, the counties in the fourth quartile for % of population that is in poverty (counties with the most poverty) receive \$ 6K to \$8K of grant funding per person. 
* In Nevada, counties in the fourth quartile receive \$ 62K - \$ 18K less grant funding per person than counties in the first quartile. 

## Table

```{r, echo=FALSE, message=FALSE, warning=FALSE}

# Getting columns together 
myAmount_table <- left_join(myAmount_map1, myAmount_map2, by = "region") %>%
  rename(perCap_1 = value.x) %>%
  rename(perCap_2 = value.y) 
myAmount_table <- left_join(myAmount_table, myAmount_map3, by = "region") %>%
  rename(perCap_3 = value) 
myAmount_table <- left_join(myAmount_table, myAmount_map4, by = "region") %>%
  rename(perCap_4 = value) 
myAmount_table <- left_join(myAmount_table, myAmount_difference_map, by = "region") %>%
  rename(perCap_diff = value) %>%
  select(-value_4, -value_1)

#get percent of poverty for the state
state_pct_poverty <- myWorking %>%
  select(region, inc_below_pov_pct) %>%
  distinct() %>%
  group_by(region) %>%
  mutate(avg_inc_below_pov_pct = mean(inc_below_pov_pct)) %>%
  select(region, avg_inc_below_pov_pct) %>%
  distinct()

#rounding two digets for all columns
myAmount_table <- left_join(myAmount_table, state_pct_poverty, by = "region") %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  filter(region != "district of columbia") %>%
  filter(region != "puerto rico") 

myAmount_table %>%
  arrange(-avg_inc_below_pov_pct) %>%
  mutate(region = str_to_title(region)) %>%
  kbl(booktabs = T, 
      caption = "Avg. grant ($)",
      col.names = c("State",
                    "First",
                    "Second", 
                    "Third",
                    "Fourth",
                    "Difference (four minus first)", 
                    "Avg % of county that's in poverty")) %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 


# Plot of avg % poverty in state and difference between quarts 
ggplot(myAmount_table, aes(x=avg_inc_below_pov_pct, y=perCap_diff)) + geom_point() + geom_smooth(method = "lm") +theme_bw() + 
  xlab("Avg. % of people in poverty per County") +
  ylab("Difference in fouth and first quartile") + ggtitle("Avg. spending")

```

When looking at the difference in the avg. funding received, there is a *very* slight downward trend with a counties' population that is in poverty. 


## Time Differences: total grant award ($)

```{r, echo=FALSE, message=FALSE, warning=FALSE}

myTime_amount <- function(myStart = 1965, myEnd = 2018){
myWorking <- myWorking %>%
    dplyr::filter(year >= myStart) %>%
    dplyr::filter(year < myEnd)
  
# First quartile 
myAmount_map1 <- myWorking %>%
  filter(pov_quarts == "1") %>%
  group_by(region) %>%
  mutate(mean_amount = mean(amount)) %>%
  select(region, mean_amount) %>%
  distinct() %>%
  rename(value = mean_amount)

# Fourth quartile 
myAmount_map4 <- myWorking %>%
  filter(pov_quarts == "4") %>%
  group_by(region) %>%
  mutate(mean_amount = mean(amount)) %>%
  select(region, mean_amount) %>%
  distinct() %>%
  rename(value = mean_amount)

# -- difference -- #
myAmount_difference_map <- left_join(x = myAmount_map4, y = myAmount_map1, by = "region") %>%
  rename(value_4 = value.x) %>%
  rename(value_1 = value.y) %>%
  mutate(value = value_4 - value_1) %>%
  filter(!is.na(region))


state_choropleth(myAmount_difference_map,
  title = paste("Difference in avg. grant between fourth quartile and first", myStart, "-", myEnd),
  legend = "Difference in $",
  zoom = continental_us_states,
  num_colors = 5) +
  scale_fill_brewer(palette="Purples", direction = -1) + 
  labs(fill = "Difference in $")


}

myTime_amount()
myTime_amount(myEnd = 1981)
myTime_amount(myStart = 1981, myEnd = 2008)
myTime_amount(myStart = 2008)


```

At first glance, it looks like the post-obama years seem to be making this trend less severe in certain states.
