==> 06.descriptive_reverse_quarts.Rmd <==
---
title: "Quartiles of Grants Described"
author: "Andie Creel"
date: "7/28/2021"
output: 
  pdf_document:
    number_sections: true
    extra_dependencies: ["float"]

knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, encoding = encoding, output_file = file.path(dirname(inputFile), 'Exploratory_Output/descriptive/grant_quarts.pdf')) })
---

Brown colors tend to indicated negative numbers and purple tend to indicate positive numbers. 

**Active years: 1975-2018 because the non-hispanic white characteristic is not available prior to 1980**


I get quartiles for the quantity of grants received after grouping by state and decade. I categorize a county with a "1" if the county receives the fewest number of grants. I categorize a county with a "4" if it receives the most amount of grants. Below, I look at the average percent of the population that is a certain demographic characteristic (ie white, people of color, below poverty line, etc) and see if there are trends in the percentage of certain demographic characteristics between the counties that get categorized with a 1 versus a 4. I do this averaging across all states and then do it again after dropping states whose counties' average population contains less than 10% POC (because they may be so homogeneous that they damped any true trends across quartiles that exist in more diverse states).


I repeat this process for the amount of funding received and the amount of funding received per capita. 



```{r, include=FALSE}
#R Markdown stuff
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(fig.pos = "!H", out.extra = "")
options(tinytex.verbose = TRUE)
```

```{r setup, include=FALSE}
library(vroom)
library(tidyr)
library(stringr)
library(usmap)
library(tidycensus)
library(ggplot2)
library(fixest)
library(kableExtra)
library(gtools)
library(forcats)
library("wesanderson")
options(scipen=99999)
library(janitor)
library(dplyr)
# ---- #
library(choroplethr)
library(choroplethrMaps)
library(usdata)
library(maps)
data(fips_codes)
data(continental_us_states)
# --- #



# made in 01.data_wrangle.R
myWorking <- vroom("Datasets/clean_data/lwcf_annualPop_deccenialDemographics.csv") 

# Grant amounts are already adjusted for inflation, adjusting median income (2018 dollars)
# additional data cleaning ------------
# Removing counties whose FIPSs code I didn't identify and don't have demo data for (1975-2018)
myWorking <- myWorking %>%
  filter(!is.na(fips)) %>%
  distinct() %>%
  mutate(type = as.factor(type)) %>%
  mutate(state_fips = as.factor(str_sub(fips, 1, 2))) %>%
  filter(!is.na(white_pct)) %>%
  dplyr::mutate(poc_pct = 100 - white_pct)
# Grouping outcome variables by decade (number of grants received, amount received, average amount received per capita per decade, avg # grants per 100,000 people per decade)
myWorking_decade <- myWorking %>%
  group_by(fips, merge_year) %>%
  mutate(decade_grants = sum(got_grant)) %>%
  mutate(decade_amount = sum(amount, na.rm = TRUE)) %>%
  mutate(decade_avg_per_cap_amount = sum(amount, na.rm = TRUE)/mean(annual_population, na.rm = TRUE)) %>%
  mutate(decade_avg_per_cap_grants = sum(got_grant)/mean((annual_population/100000), na.rm = TRUE)) %>%
  select(state_fips, ends_with("_pct"), starts_with("decade_")) %>%
  ungroup() %>%
  distinct()
#getting average POC per county
state_pct_POC <- myWorking_decade %>%
  select(state_fips, poc_pct) %>%
  distinct() %>%
  group_by(state_fips) %>%
  mutate(avg_poc_pct = mean(poc_pct)) %>%
  select(state_fips, avg_poc_pct) %>%
  distinct()
myWorking_decade <- left_join(myWorking_decade, state_pct_POC, by = "state_fips" )
# ------- Get quartiles of grants received for each decade ------- #
myQuants <- myWorking_decade %>%
  dplyr::group_by(state_fips, merge_year) %>%
  dplyr::summarise(quants = list(quantile(decade_grants, probs = c(.25, .5, .75)))) %>%
  unnest_wider(quants) %>%
  dplyr::rename(decade_grants_25 = `25%`) %>%
  dplyr::rename(decade_grants_50 = `50%`) %>%
  dplyr::rename(decade_grants_75 = `75%`)
#joining back in
myWorking_decade <- left_join(myWorking_decade, myQuants, by = c("state_fips", "merge_year"))
rm(myQuants)
#making a factor variable for the quartiles (by state and year)
myWorking_decade <- myWorking_decade %>%
  dplyr::mutate(decade_grants_quants = if_else(decade_grants > decade_grants_75, "4", "")) %>% # fourth quartile 
  dplyr::mutate(decade_grants_quants = if_else(decade_grants <=decade_grants_75, "3", decade_grants_quants)) %>% # third quartile
  dplyr::mutate(decade_grants_quants = if_else(decade_grants <= decade_grants_50, "2", decade_grants_quants)) %>% # second quartile
  dplyr::mutate(decade_grants_quants = if_else(decade_grants <= decade_grants_25, "1", decade_grants_quants)) %>% # first quartile 
  dplyr::select(-decade_grants_75, -decade_grants_50, - decade_grants_25) %>%
  dplyr::mutate(decade_grants_quants = as.factor(decade_grants_quants))
# ------- Get quartiles of amount of grant funding for each decade ------- #
myQuants <- myWorking_decade %>%
  dplyr::group_by(state_fips, merge_year) %>%
  dplyr::summarise(quants = list(quantile(decade_amount, probs = c(.25, .5, .75)))) %>%
  unnest_wider(quants) %>%
  dplyr::rename(decade_amount_25 = `25%`) %>%
  dplyr::rename(decade_amount_50 = `50%`) %>%
  dplyr::rename(decade_amount_75 = `75%`)
#joining back in
myWorking_decade <- left_join(myWorking_decade, myQuants, by = c("state_fips", "merge_year"))
rm(myQuants)
#making a factor variable for the quartiles (by state and year)
myWorking_decade <- myWorking_decade %>%
  dplyr::mutate(decade_amount_quants = if_else(decade_amount > decade_amount_75, "4", "")) %>% # fourth quartile 
  dplyr::mutate(decade_amount_quants = if_else(decade_amount <=decade_amount_75, "3", decade_amount_quants)) %>% # third quartile
  dplyr::mutate(decade_amount_quants = if_else(decade_amount <= decade_amount_50, "2", decade_amount_quants)) %>% # second quartile
  dplyr::mutate(decade_amount_quants = if_else(decade_amount <= decade_amount_25, "1", decade_amount_quants)) %>% # first quartile 
  dplyr::select(-decade_amount_75, -decade_amount_50, - decade_amount_25) %>%
  dplyr::mutate(decade_amount_quants = as.factor(decade_amount_quants))
# ------- Get quartiles of amount of grant funding PER CAPITA for each decade ------- #
myQuants <- myWorking_decade %>%
  dplyr::group_by(state_fips, merge_year) %>%
  dplyr::summarise(quants = list(quantile(decade_avg_per_cap_amount, probs = c(.25, .5, .75), na.rm = TRUE))) %>%
  unnest_wider(quants) %>%
  dplyr::rename(decade_avg_per_cap_amount_25 = `25%`) %>%
  dplyr::rename(decade_avg_per_cap_amount_50 = `50%`) %>%
  dplyr::rename(decade_avg_per_cap_amount_75 = `75%`)
#joining back in
myWorking_decade <- left_join(myWorking_decade, myQuants, by = c("state_fips", "merge_year"))
rm(myQuants)
#making a factor variable for the quartiles (by state and year)
myWorking_decade <- myWorking_decade %>%
  dplyr::mutate(decade_avg_per_cap_amount_quants = if_else(decade_avg_per_cap_amount > decade_avg_per_cap_amount_75, "4", "")) %>% # fourth quartile 
  dplyr::mutate(decade_avg_per_cap_amount_quants = if_else(decade_avg_per_cap_amount <=decade_avg_per_cap_amount_75, "3", decade_avg_per_cap_amount_quants)) %>% # third quartile
  dplyr::mutate(decade_avg_per_cap_amount_quants = if_else(decade_avg_per_cap_amount <= decade_avg_per_cap_amount_50, "2", decade_avg_per_cap_amount_quants)) %>% # second quartile
  dplyr::mutate(decade_avg_per_cap_amount_quants = if_else(decade_avg_per_cap_amount <= decade_avg_per_cap_amount_25, "1", decade_avg_per_cap_amount_quants)) %>% # first quartile 
  dplyr::select(-decade_avg_per_cap_amount_75, -decade_avg_per_cap_amount_50, - decade_avg_per_cap_amount_25) %>%
  dplyr::mutate(decade_avg_per_cap_amount_quants = as.factor(decade_avg_per_cap_amount_quants))
# ------- Get quartiles of number of grants  PER CAPITA for each decade ------- #
myQuants <- myWorking_decade %>%
  dplyr::group_by(state_fips, merge_year) %>%
  dplyr::summarise(quants = list(quantile(decade_avg_per_cap_grants, probs = c(.25, .5, .75), na.rm = TRUE))) %>%
  unnest_wider(quants) %>%
  dplyr::rename(decade_avg_per_cap_grants_25 = `25%`) %>%
  dplyr::rename(decade_avg_per_cap_grants_50 = `50%`) %>%
  dplyr::rename(decade_avg_per_cap_grants_75 = `75%`)
#joining back in
myWorking_decade <- left_join(myWorking_decade, myQuants, by = c("state_fips", "merge_year"))
rm(myQuants)
#making a factor variable for the quartiles (by state and year)
myWorking_decade <- myWorking_decade %>%
  dplyr::mutate(decade_avg_per_cap_grants_quants = if_else(decade_avg_per_cap_grants > decade_avg_per_cap_grants_75, "4", "")) %>% # fourth quartile 
  dplyr::mutate(decade_avg_per_cap_grants_quants = if_else(decade_avg_per_cap_grants <=decade_avg_per_cap_grants_75, "3", decade_avg_per_cap_grants_quants)) %>% # third quartile
  dplyr::mutate(decade_avg_per_cap_grants_quants = if_else(decade_avg_per_cap_grants <= decade_avg_per_cap_grants_50, "2", decade_avg_per_cap_grants_quants)) %>% # second quartile
  dplyr::mutate(decade_avg_per_cap_grants_quants = if_else(decade_avg_per_cap_grants <= decade_avg_per_cap_grants_25, "1", decade_avg_per_cap_grants_quants)) %>% # first quartile 
  dplyr::select(-decade_avg_per_cap_grants_75, -decade_avg_per_cap_grants_50, - decade_avg_per_cap_grants_25) %>%
  dplyr::mutate(decade_avg_per_cap_grants_quants = as.factor(decade_avg_per_cap_grants_quants))

```

# Maps

## Quantity of grants per capita 


```{r, echo=FALSE, message=FALSE, warning=FALSE}
# QUANTITY OF GRANTS PER CAPITA
# fourth quartile is receiving most amount of grants 
# difference in % POC from 1st quart and 4th quart -------------
myMap_poc <- myWorking_decade %>%
  select(state_fips, fips, merge_year, poc_pct, decade_avg_per_cap_grants_quants) %>%
  pivot_wider(names_from = decade_avg_per_cap_grants_quants, values_from = c(poc_pct)) %>%
  group_by(state_fips) %>%
  mutate(poc_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(poc_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(poc_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(poc_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("poc_pct_")) %>%
  distinct()
#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()
myMap_poc <- left_join(myMap_poc, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(poc_pct_4 - poc_pct_1), 2)
state_choropleth(myMap_poc,
  zoom = continental_us_states) +
  # scale_fill_brewer(palette="YlOrBr", direction = -1)+
  scale_fill_brewer(palette="PuOr")+
  ggtitle(label = "POC: Top quartile for # grants per capita vs the bottom quartile", 
          subtitle =  "Difference in avg % of POC (top minus bottom)") + 
  labs(fill = "Difference in %") +
  theme(legend.position="bottom")
# difference in % poverty from 1st quart and 4th quart -------------
myMap_pov <- myWorking_decade %>%
  select(state_fips, fips, merge_year, inc_below_pov_pct, decade_avg_per_cap_grants_quants) %>%
  pivot_wider(names_from = decade_avg_per_cap_grants_quants, values_from = c(inc_below_pov_pct)) %>%
  group_by(state_fips) %>%
  mutate(inc_below_pov_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("inc_below_pov_pct_")) %>%
  distinct()
#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()
myMap_pov <- left_join(myMap_pov, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(inc_below_pov_pct_4 - inc_below_pov_pct_1), 2)
state_choropleth(myMap_pov,
  zoom = continental_us_states) +
  # scale_fill_brewer(palette="YlOrBr", direction = -1)+
  scale_fill_brewer(palette="PuOr")+
  ggtitle(label = "Poverty: Top quartile for # grants per capita vs the bottom quartile", 
          subtitle =  "Difference in avg % in poverty (top minus bottom)") + 
  labs(fill = "Difference in % ") +
  theme(legend.position="bottom")
# difference in % urban from 1st quart and 4th quart -------------
myMap_rural <- myWorking_decade %>%
  select(state_fips, fips, merge_year, rural_pct, decade_avg_per_cap_grants_quants) %>%
  pivot_wider(names_from = decade_avg_per_cap_grants_quants, values_from = c(rural_pct)) %>%
  group_by(state_fips) %>%
  mutate(rural_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(rural_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(rural_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(rural_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("rural_pct_")) %>%
  distinct()
#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()
myMap_rural <- left_join(myMap_rural, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(rural_pct_4 - rural_pct_1), 2)
state_choropleth(myMap_rural,
  zoom = continental_us_states) +
  scale_fill_brewer(palette="PuOr")+
  ggtitle(label = "Rural: Top quartile for # grants per capita vs the bottom quartile", 
          subtitle =  "Difference in avg % rural (top minus bottom)") + 
  labs(fill = "Difference in %") +
  theme(legend.position="bottom")


```
Interpretation: 
In MS, AL, and GA, the counties that receive the most number grants per capita in a given decade have **6-14% less POC** than counties that receive the least number of grants per capita.

In WV, the counties that receive the most number grants per capita in a given decade have **1-4% less people in poverty** than counties that receive the least number of grants per capita.

In TX, the counties that receive the most number grants per capita in a given decade have **6- 21% less rural people** than counties that receive the least number of grants per capita.

## Amount of grant funding per capita 


```{r, echo=FALSE, message=FALSE, warning= FALSE}
rm(myMap_poc, myMap_pov, myMap_rural)

# AMOUNT OF GRANT $$ PER CAPITA

# difference in % POC from 1st quart and 4th quart -------------

myMap_poc <- myWorking_decade %>%
  select(state_fips, fips, merge_year, poc_pct, decade_avg_per_cap_amount_quants) %>%
  pivot_wider(names_from = decade_avg_per_cap_amount_quants, values_from = c(poc_pct)) %>%
  group_by(state_fips) %>%
  mutate(poc_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(poc_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(poc_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(poc_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("poc_pct_")) %>%
  distinct()

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myMap_poc <- left_join(myMap_poc, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(poc_pct_4 - poc_pct_1), 2)

state_choropleth(myMap_poc,
  zoom = continental_us_states) +
  # scale_fill_brewer(palette="YlOrBr", direction = -1)+
  scale_fill_brewer(palette="PuOr")+
  ggtitle(label = "POC: Top quartile for amount of funding per capita vs the bottom quartile", 
          subtitle =  "Difference in avg % of POC (top minus bottom)") + 
  labs(fill = "Difference in % ")+
  theme(legend.position="bottom")


 # difference in % poverty from 1st quart and 4th quart -------------

myMap_pov <- myWorking_decade %>%
  select(state_fips, fips, merge_year, inc_below_pov_pct, decade_avg_per_cap_amount_quants) %>%
  pivot_wider(names_from = decade_avg_per_cap_amount_quants, values_from = c(inc_below_pov_pct)) %>%
  group_by(state_fips) %>%
  mutate(inc_below_pov_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("inc_below_pov_pct_")) %>%
  distinct()

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myMap_pov <- left_join(myMap_pov, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(inc_below_pov_pct_4 - inc_below_pov_pct_1), 2)

state_choropleth(myMap_pov,
  zoom = continental_us_states) +
  scale_fill_brewer(palette="PuOr")+
  # scale_fill_brewer(palette="Purples")+
  ggtitle(label = "Poverty: Top quartile for amount of funding per capita vs the bottom quartile", 
          subtitle =  "Difference in avg % in poverty (top minus bottom)") + 
  labs(fill = "Difference in % ")+
  theme(legend.position="bottom")


# difference in % urban from 1st quart and 4th quart -------------

myMap_rural <- myWorking_decade %>%
  select(state_fips, fips, merge_year, rural_pct, decade_avg_per_cap_amount_quants) %>%
  pivot_wider(names_from = decade_avg_per_cap_amount_quants, values_from = c(rural_pct)) %>%
  group_by(state_fips) %>%
  mutate(rural_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(rural_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(rural_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(rural_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("rural_pct_")) %>%
  distinct()

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myMap_rural <- left_join(myMap_rural, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(rural_pct_4 - rural_pct_1), 2)

state_choropleth(myMap_rural,
  zoom = continental_us_states) +
  # scale_fill_brewer(palette="Purples")+
  scale_fill_brewer(palette="PuOr")+
  ggtitle(label = "Rural: Top quartile for amount of funding per capita vs the bottom quartile", 
          subtitle =  "Difference in avg % rural (top minus bottom)") + 
  labs(fill = "Difference in % ")+
  theme(legend.position="bottom")


```
Interpretation: 
In California, the counties that receive the most amount of grant funding per capita in a given decade have **9-22% less POC** than counties that receive the least amount of grant funding per capita.

In LA, the counties that receive the most amount of grant funding per capita in a given decade have **4-6% more people in poverty** than counties that receive the least amount of grant funding per capita.

In every state, the counties that receive the most most amount of grant funding per capita in a given decade have **more rural people** than counties that receive the least amount of grant funding per capita.


*Concern: potential for the trend we're seeing in POC receiving less grants could just be driven by POC living in urban places.* 
However, I think it's still valid for us to look at total quantity and total amount funding because parks are a public good, meaning their less exclusionary. I think if use was more exclusionary, the per capita measure would be really important. But for a public good, multiple people can use the park over and over again. Per capita is an important metric but not as critical in this case as it may be in others. 

## Quantity of grants


```{r, echo=FALSE, message=FALSE, warning=FALSE}
# QUANTITY OF GRANTS 

# fourth quartile is receiving most amount of grants 

# difference in % POC from 1st quart and 4th quart -------------

myMap_poc <- myWorking_decade %>%
  select(state_fips, fips, merge_year, poc_pct, decade_grants_quants) %>%
  pivot_wider(names_from = decade_grants_quants, values_from = c(poc_pct)) %>%
  group_by(state_fips) %>%
  mutate(poc_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(poc_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(poc_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(poc_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("poc_pct_")) %>%
  distinct()

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myMap_poc <- left_join(myMap_poc, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(poc_pct_4 - poc_pct_1), 2)

state_choropleth(myMap_poc,
  zoom = continental_us_states) +
  # scale_fill_brewer(palette="YlOrBr", direction = -1)+
  scale_fill_brewer(palette="PuOr")+
  ggtitle(label = "POC: Top quartile for # grants vs the bottom quartile", 
          subtitle =  "Difference in avg % of POC (top minus bottom)") + 
  labs(fill = "Difference in %") +
  theme(legend.position="bottom")


# difference in % poverty from 1st quart and 4th quart -------------

myMap_pov <- myWorking_decade %>%
  select(state_fips, fips, merge_year, inc_below_pov_pct, decade_grants_quants) %>%
  pivot_wider(names_from = decade_grants_quants, values_from = c(inc_below_pov_pct)) %>%
  group_by(state_fips) %>%
  mutate(inc_below_pov_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("inc_below_pov_pct_")) %>%
  distinct()

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myMap_pov <- left_join(myMap_pov, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(inc_below_pov_pct_4 - inc_below_pov_pct_1), 2)

state_choropleth(myMap_pov,
  zoom = continental_us_states) +
  # scale_fill_brewer(palette="YlOrBr", direction = -1)+
  scale_fill_brewer(palette="PuOr")+
  ggtitle(label = "Poverty: Top quartile for # grants vs the bottom quartile", 
          subtitle =  "Difference in avg % in poverty (top minus bottom)") + 
  labs(fill = "Difference in % ") +
  theme(legend.position="bottom")


# difference in % urban from 1st quart and 4th quart -------------

myMap_rural <- myWorking_decade %>%
  select(state_fips, fips, merge_year, rural_pct, decade_grants_quants) %>%
  pivot_wider(names_from = decade_grants_quants, values_from = c(rural_pct)) %>%
  group_by(state_fips) %>%
  mutate(rural_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(rural_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(rural_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(rural_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("rural_pct_")) %>%
  distinct()

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myMap_rural <- left_join(myMap_rural, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(rural_pct_4 - rural_pct_1), 2)

state_choropleth(myMap_rural,
  zoom = continental_us_states) +
  # scale_fill_brewer(palette="PuOr")+
  scale_fill_brewer(palette="YlOrBr", direction = -1)+
  ggtitle(label = "Rural: Top quartile for # grants vs the bottom quartile", 
          subtitle =  "Difference in avg % rural (top minus bottom)") + 
  labs(fill = "Difference in %") +
  theme(legend.position="bottom")


```
Interpretation: 
In MS, AL, GA and SC, the counties that receive the highest quantity of grants in a given decade have **6-22% less POC** than counties that receive the smallest quantity of grants.

In MS and AL, the counties that receive the highest quantity of grants in a given decade have **4-11% less people in poverty** than counties that receive the smallest quantity of grants.

For the majority of the country, the counties that receive the highest quantity of grants in a given decade have **are less rural** than counties that receive the smallest quantity of grants.


## Amount of grant funding


```{r, echo=FALSE, message=FALSE, warning= FALSE}
rm(myMap_poc, myMap_pov, myMap_rural)

# AMOUNT OF GRANT $$

# difference in % POC from 1st quart and 4th quart -------------

myMap_poc <- myWorking_decade %>%
  select(state_fips, fips, merge_year, poc_pct, decade_amount_quants) %>%
  pivot_wider(names_from = decade_amount_quants, values_from = c(poc_pct)) %>%
  group_by(state_fips) %>%
  mutate(poc_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(poc_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(poc_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(poc_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("poc_pct_")) %>%
  distinct()

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myMap_poc <- left_join(myMap_poc, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(poc_pct_4 - poc_pct_1), 2)

state_choropleth(myMap_poc,
  zoom = continental_us_states) +
  # scale_fill_brewer(palette="YlOrBr", direction = -1)+
  scale_fill_brewer(palette="PuOr")+
  ggtitle(label = "POC: Top quartile for amount of funding vs the bottom quartile", 
          subtitle =  "Difference in avg % of POC (top minus bottom)") + 
  labs(fill = "Difference in % ")+
  theme(legend.position="bottom")


 # difference in % poverty from 1st quart and 4th quart -------------

myMap_pov <- myWorking_decade %>%
  select(state_fips, fips, merge_year, inc_below_pov_pct, decade_amount_quants) %>%
  pivot_wider(names_from = decade_amount_quants, values_from = c(inc_below_pov_pct)) %>%
  group_by(state_fips) %>%
  mutate(inc_below_pov_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(inc_below_pov_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("inc_below_pov_pct_")) %>%
  distinct()

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myMap_pov <- left_join(myMap_pov, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(inc_below_pov_pct_4 - inc_below_pov_pct_1), 2)

state_choropleth(myMap_pov,
  zoom = continental_us_states,
  num_colors = 7) +
  # scale_fill_brewer(palette="PuOr")+
  scale_fill_brewer(palette="YlOrBr", direction = -1)+
  # scale_fill_brewer(palette="Purples")+
  ggtitle(label = "Poverty: Top quartile for amount of funding vs the bottom quartile", 
          subtitle =  "Difference in avg % in poverty (top minus bottom)") + 
  labs(fill = "Difference in % ")+
  theme(legend.position="bottom")


# difference in % urban from 1st quart and 4th quart -------------

myMap_rural <- myWorking_decade %>%
  select(state_fips, fips, merge_year, rural_pct, decade_amount_quants) %>%
  pivot_wider(names_from = decade_amount_quants, values_from = c(rural_pct)) %>%
  group_by(state_fips) %>%
  mutate(rural_pct_4 = mean(`4`, na.rm = TRUE)) %>%
  mutate(rural_pct_3 = mean(`3`, na.rm = TRUE)) %>%
  mutate(rural_pct_2 = mean(`2`, na.rm = TRUE)) %>%
  mutate(rural_pct_1 = mean(`1`, na.rm = TRUE)) %>%
  select(state_fips, starts_with("rural_pct_")) %>%
  distinct()

#getting state name for choropleth maps 
fips_codes <- fips_codes %>%
  select(state_code, state_name) %>%
  distinct()

myMap_rural <- left_join(myMap_rural, fips_codes, by = c("state_fips" = "state_code")) %>% 
  rename(region = state_name) %>%
  mutate(region = tolower(region)) %>%
  mutate(value = round(rural_pct_4 - rural_pct_1), 2)

state_choropleth(myMap_rural,
  zoom = continental_us_states) +
  scale_fill_brewer(palette="YlOrBr", direction = -1)+
  # scale_fill_brewer(palette="Purples")+
  ggtitle(label = "Rural: Top quartile for amount of funding vs the bottom quartile", 
          subtitle =  "Difference in avg % rural (top minus bottom)") + 
  labs(fill = "Difference in % ")+
  theme(legend.position="bottom")


```
In MS, AL, GA and SC, the counties that receive the highest amount of grant funding in a given decade have **5-22% less POC** than counties that receive the smallest qamount of grant funding.

In MS, FL and AL, the counties that receive the highest amount of grant funding in a given decade have **4-10% less people in poverty** than counties that receive the lowest amount of grant funding.

For the majority of the county, the counties that receive the highest amount of grant fundin in a given decade have **are less rural** than counties that receive the smallest amount of grant funding.
