==> 09.suits_index.Rmd <==
---
title: "Suits Index"
author: "Andie Creel"
date: 'September, 2022'
output: 
  html_document:
    number_sections: true
---

```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list = ls())
library(vroom)
library(tidyr)
library(stringr)
library(ggplot2)
library(gtools)
library(forcats)
options(scipen=99999)
library(janitor)
library(dplyr)
library(zoo)
library(runner)
source('09.functions.R')
library(fixest)
library(kableExtra)
library(ggpubr)

# ----maps ---- #
library(choroplethr)
library(choroplethrMaps)
library(usdata)
library(maps)
library(tidycensus)
data("fips_codes")
data("state.regions")
library(RColorBrewer)


```


Thanks to Matt Ashenfarb for providing helpful code on the suites index. 
```{r, include=FALSE, message=FALSE}
load("09.suits_environment.RData")
```

# Set Up and Definitions 

The **outcomes of interest** is how the average (through time) LWCF dollars per capita distributed in each county. I look at the distribution in average LWCF dollars per capita (investment per capita) across different EJ criteria.  

Originally I used two different per capita measurements for the amount of LWCF funds.

1. Funds given to a county  divided by the population of the county at the time of LWCF distribution. This captures how the program was administered through time.
2. The cumulative amount of funds given to a county from 1965-2018 divided by the population of the county in 2020 (estimates for population of each county is not available for each county in 2018 in the American Community Survey). This captures how individuals today are effected by the LWCF program.

After both per capita metrics returned similar results (stored in commit from Sept 20th, 2022), I used the first definition which can be interpreted as how the program has been administered historically. Investment per capita is the average of what is described in one. 

I use per capita metrics for amount of funding to control for a county's population. 

## Demographics of Interest (x axis)
We are interested in the distribution of funds across three county-level demographic characteristics: median income, % POC, % rural. 

**Average median household income**: the median household income (averaged from 1975-2018 for each county bc median income was not collected in the 1970 census). 

**Average percent people of color**: the percent of people in a color who are not non-Hispanic white (averaged from 1965-2018 for each county). 

**Average percent rural**: the percent of people living in a rural cbg. (averaged from 1965-2018 for each county). 


## Comparing of Suits Index across states 
We can compare the level of disparity (i.e., suits index) across states. The level of disparity can be compared across states when thinking about how a state treats it's **percentiles**, however it's important to bare in mind that the county in Montana with the most POC will still be far whiter than many "average" counties in the south. 

Additionally, two states with different income distributions could have the same suits index. Two may have different income distributions (i.e., Michigan is on average richer than New Mexico) but have the same suits index.  

## The sign of the suits index 
Suits indices are traditionally used to evaluate if a **tax** is progressive or regressive. A positive suits index is a progressive tax. The suits index is **typically** calculated as [(area under the neutral line) - (area under Lorenz curve)] / (area under the neutral line). 

The LWCF is a **subsidy** program. To keep signs consistent with "progressive" (favor an EJ group) or "regressive" (Disfavors an EJ group) **I calculated the suits index as [(area under  Lorenz curve) - (area under the neutral line)] / (area under the neutral line)**.

**A negative suits index in our study indicates EJ communities receive less than their proportional share .**

# National Results

```{r, echo=FALSE, message=FALSE}
#-------------------------------------------------------------------------------
# National results for income
#-------------------------------------------------------------------------------

myResult_income <- getSuits(myWorking, x_axis = "Income")
national_income <- getGraph(myResult_income, x_axis = "Income", caption = "Investment per capita is distributed proportionally across counties with varying average median income. ") 

ggsave(filename ="Figures/national_median_income.jpeg", plot = national_income,  width = 11, height = 8)

#-------------------------------------------------------------------------------
# National results for POC
#-------------------------------------------------------------------------------

myResult_poc <- getSuits(myWorking, x_axis = "POC")
national_poc <- getGraph(myResult_poc, x_axis = "POC")

ggsave(filename ="Figures/national_POC.jpeg", plot = national_poc, width = 11, height = 8)


#-------------------------------------------------------------------------------
# National results for Rural
#-------------------------------------------------------------------------------

myResult_rural <- getSuits(myWorking, x_axis = "Rural")
national_rural <- getGraph(myResult_rural, x_axis = "Rural")
ggsave(filename ="Figures/national_rural.jpeg", plot = national_rural, width = 11, height = 8)

#-------------------------------------------------------------------------------
# make figure with three graphs -- national 
#-------------------------------------------------------------------------------
tripple_national <- ggarrange(national_income, national_poc, national_rural,
                    labels = c("Median Income", "People of Color", "Rurality"),
                    label.x = c(0,0,.09),
                    # hjust = -.75,
                    vjust = 3,
                    ncol = 3, nrow = 1) 

ggsave(filename ="Figures/national_all.jpeg", plot = tripple_national, width = 12, height = 4)

tripple_national


```

# State Results

```{r, echo=FALSE}
#rewrote the state_choropleth function a bit to get rid of state label
# NOTE: num_colors = 0 makes it so white is zero

#-------------------------------------------------------------------------------
# State Results for income
#-------------------------------------------------------------------------------
state_income <- myState_choropleth(myMap_income, num_colors = 0, legend = "Suits Index" ) 
  # ggtitle(label = "Suits indexes for income varries state to state despite being proportional nationally") 
  
ggsave(filename ="Figures/state_median_income.jpeg", plot = state_income, width = 11, height = 8)

#-------------------------------------------------------------------------------
# State Results for POC
#-------------------------------------------------------------------------------
state_poc <- myState_choropleth(myMap_poc, num_colors = 0, legend = "Suits Index" ) 
  # ggtitle(label = "All states disfavor counties with more people of color")
  # theme(legend.position="bottom")

ggsave(filename ="Figures/state_poc.jpeg", plot = state_poc, width = 11, height = 8)

#-------------------------------------------------------------------------------
# State Results for Rural
#-------------------------------------------------------------------------------
state_rural <- myState_choropleth(myMap_rural, num_colors = 0, legend = "Suits Index" ) 
  # ggtitle(label = "Rural counties are disfavored",
          # subtitle =  "However, not to the extent counties with many POC are") 
  # theme(legend.position="bottom")

ggsave(filename ="Figures/state_rural.jpeg", plot = state_rural, width = 11, height = 8)

#-------------------------------------------------------------------------------
# make figure with three graphs -- state 
#-------------------------------------------------------------------------------
tripple_state <- ggarrange(state_income, state_poc, state_rural,
                    labels = c("Poverty", "People of Color", "Rurality"),
                    # label.x = c(0,0,.09),
                    hjust = c(-.1,-.1, -.2),
                    vjust = 8,
                    ncol = 3, nrow = 1, 
                    common.legend = TRUE, 
                    legend = "right") 

ggsave(filename ="Figures/state_all.jpeg", plot = tripple_state, width = 12, height = 4)

tripple_state

```


# Regressions 

## Income 
```{r, echo = FALSE}
#-------------------------------------------------------------------------------
# regressions with the exact same dataset used for the suits indices
#-------------------------------------------------------------------------------
one <- feols(data = myWorking, amount_per_cap_avg ~  avg_med_income_house)
two <- feols(data = myWorking, amount_per_cap_avg ~  avg_med_income_house + avg_poc_pct)
three <- feols(data = myWorking, amount_per_cap_avg ~  avg_med_income_house + avg_poc_pct + avg_rural_pct)
four <- feols(data = myWorking, amount_per_cap_avg ~  avg_poc_pct + avg_rural_pct + avg_med_income_house | state_fips)

etable(one,
       two,
       three,
       four,
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

```

Income coefficient is stable across regression


## Population of Color
```{r, echo = FALSE}
#-------------------------------------------------------------------------------
# regressions with the exact same dataset used for the suits indices
#-------------------------------------------------------------------------------
one <- feols(data = myWorking, amount_per_cap_avg ~  avg_poc_pct)
two <- feols(data = myWorking, amount_per_cap_avg ~  avg_poc_pct + avg_med_income_house)
three <- feols(data = myWorking, amount_per_cap_avg ~  avg_poc_pct + avg_rural_pct + avg_med_income_house)
four <- feols(data = myWorking, amount_per_cap_avg ~  avg_poc_pct + avg_rural_pct + avg_med_income_house | state_fips)

etable(one,
       two,
       three,
       four,
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 


avg_amount_per_cap_avg <- round(mean(myWorking$amount_per_cap_avg, na.rm = TRUE),2)
```

POC coefficient is fairly stable across regression (consistently negative, economically significant and statistically significant). The average LWCF dollars per capita is $`r avg_amount_per_cap_avg`, so this indicates ~1.5% less funding for a 1% increase in POC. 


## Rural 
```{r, echo = FALSE}
#-------------------------------------------------------------------------------
# regressions with the exact same dataset used for the suits indices
#-------------------------------------------------------------------------------
one <- feols(data = myWorking, amount_per_cap_avg ~  avg_rural_pct)
two <- feols(data = myWorking, amount_per_cap_avg ~  avg_rural_pct + avg_med_income_house)
three <- feols(data = myWorking, amount_per_cap_avg ~  avg_poc_pct + avg_rural_pct + avg_med_income_house)
four <- feols(data = myWorking, amount_per_cap_avg ~  avg_poc_pct + avg_rural_pct + avg_med_income_house | state_fips)

etable(one,
       two,
       three,
       four,
       signifCode = c("***"=0.01, "**"=0.05, "*"=0.10)) %>%
  kbl(booktabs = T, caption = "Race") %>%
  kable_styling(latex_options = c("scale_down","HOLD_position")) 

```

Rural coefficient is not very stable across regressions (indicating an endogenous issue).  


Old comment: This could be driven by states with more urban areas getting more funds, but then within those states their relatively rural counties get less funding (look at state map)s

# Conclusions 

* LWCF seems fairly proportional for income and wealth
* LWCF  disfavors counties with more POC. This matches up with what we saw in the quartile study, and what Katharine Sims has found.
* There is a less clear story for rural vs urban counties. I think rural status is so linked with population, which is correlated with the amount of funding the state gets from the federal program. It's harder to identify the effect here.  

# Remaining to-dos:
* do bootstrap-t confidence intervals for suits indices 
* look at obama admin onward of funds to see if trend has changed in last decade
* merge in 2020 census data instead ACS
