==> 10.data_sum_stats.Rmd <==
---
title: "Data Summary Stats"
author: "Andie Creel"
date: "2023-01-29"
output:
  pdf_document:
    keep_tex: yes
    latex_engine: pdflatex
    fig_caption: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list = ls())
library(vroom)
library(tidyr)
library(stringr)
library(ggplot2)
library(gtools)
library(forcats)
options(scipen=99999)
library(janitor)
library(dplyr)
 library(fixest)
library(kableExtra)
library(stargazer)


```

## Summary Statistics about our dataset 

### How many individual grants did TWS have data on?

```{r, results='asis', message=FALSE}
# made in 01.data_wrangle.R
myTWS <- vroom("Datasets/clean_data/lwcf_annualPop_deccenialDemographics.csv") 

myTWS <- myTWS %>% 
  select(fips, year, type, amount) %>%
  filter(!is.na(type)) %>%
  filter(amount != 0) %>%
  distinct()

tws_grant_num <- length(myTWS$type)


type_percents <- count(myTWS, type) %>%
  mutate(perc = round(n/tws_grant_num, 2))

type_percents
```

TWS had data on `r tws_grant_num` grants where they knew the grant type, the amount of funding reward, and the county that the project was located inside of.

Types of LWCF projects 

* A: Land Acquisition
* D: Development 
* R: Redevelopment 
* C: Combination
* P: Planning

### How many individual grants stay in our dataset (where we know median income, poc % and rural %)

```{r, results='asis', message=FALSE}
load("09.suits_environment.RData")
rm(myMap_income, myMap_poc, myMap_rural, myResult_state, state_dfs)

# _clean is the orginal dataset cleaned to relevant variables
my_grant_num <- length(myWorking_clean$state_fips)

```

Our data set has `r my_grant_num` grants. 



### Summary Table

```{r, results='asis'}
makeSummaryTable <- function(df){
  
  # getting started with type break down
  mySumTable <- count(df, type) %>%
    mutate(perc = round(n/length(df$type), 2))
  
  # getting average grant size by type
  myAvg_grant_size <-  df %>%
    group_by(type) %>%
    mutate(avg_amount = mean(amount)) %>%
    select(type, avg_amount) %>%
    distinct()
    
  mySumTable <- left_join(mySumTable, myAvg_grant_size, by = "type")
  
  #creating row for total
  mySumTable[5, 1] <- "n"
  mySumTable[5, 2] <- sum(mySumTable$n, na.rm = T)
  mySumTable[5, 3] <- sum(mySumTable$perc, na.rm = T)
  mySumTable[5, 4] <- mean(df$amount)
  
  # cleaning up table 
  mySumTable <- mySumTable %>%
    mutate(avg_amount = round(avg_amount, 2)) %>%
    mutate(perc = perc*100)
  
  # replacing type 
  mySumTable$type[1] <- "Acquisition"
  mySumTable$type[2] <- "Combination"
  mySumTable$type[3] <- "Development"
  mySumTable$type[4] <- "Redevelopment"
  
  # renaming column names 
  colnames(mySumTable) <- c("Type", "Count", "Perc.", "Avg. Amount (2018$)")
  
  return(mySumTable)

}

# _clean is the orginal dataset cleaned to relevant variables
mySumTable <- makeSummaryTable(myWorking_clean)

stargazer(mySumTable, type = "latex", summary = F, rownames = F, title = "Summary of grants included in our dataset")

mySumTable

```

### what is the average investment per capita ? 


```{r}
# myWorking have average investment per capita (not long by time)
invest_per_cap_dataset <- round(mean(myWorking$amount_per_cap_avg),2)
num_counties <- length(myWorking$fips)
```

The average investment per capita is `r invest_per_cap_dataset`. We have data for `r num_counties` counties.


### chloroplast map of investment per capita in each county 

```{r}
# investigating format I need region and value in to use choropleth functions
# data(county.map)

# making dataset
myCounty_map_df <- myWorking %>%
  mutate(region = as.numeric(fips)) %>%
  mutate(value = round(amount_per_cap_avg, 2)) %>%
  select(region, value) 

# making plot
 myCounty_map <- county_choropleth(myCounty_map_df, num_colors = 9, legend = "Investment Per Capita")  + 
  theme(plot.caption.position = "plot",
        plot.caption = element_text(hjust = 0, size = 25))

ggsave(filename ="Figures/county_invest_per_cap.jpeg", myCounty_map, width = 11, height = 8)

myCounty_map
```
Figure XX: There is a lot of within state variation in investment per capita on the East Coast, South, Midwest and the West Coast. The West has less varition, and counties benefit form a higher investment per capita. Maine and Vermont follow the same pattern as the West.
